maybe midway (change to components instead of plain javascript)
