+++ src/components/MapViewer.jsx
@@ -1,0 +1,279 @@
+import React, { useState, useEffect } from "react";
+import "../App.css";
+
+
+function zoomAndPan(mainSVG, pathElement, viewBoxInitValue) {
+    let bboxPath = pathElement.getBoundingClientRect();
+    let bboxSvg = mainSVG.getBoundingClientRect();
+
+    let xCountry = bboxPath.x;
+    let yCountry = bboxPath.y;
+    let xMap = bboxSvg.x;
+    let yMap = bboxSvg.y;
+    let widthCountry = bboxPath.width;
+    let heightCountry = bboxPath.height;
+
+    let xCenterCountry = ((xCountry - xMap) + (widthCountry / 2));
+    let yCenterCountry = ((yCountry - yMap) + (heightCountry / 2));
+
+    let xCenter = xCenterCountry - (widthCountry / 2);
+    let yCenter = yCenterCountry - (heightCountry / 2);
+
+    let zoomedIn = false;
+
+    if (viewBoxInitValue !== mainSVG.getAttribute('viewBox')) {
+        mainSVG.setAttribute("viewBox", viewBoxInitValue);
+        zoomedIn = false;
+    } else {
+        mainSVG.setAttribute("viewBox", xCenter + " " + yCenter + " " + widthCountry + " " + heightCountry);
+        zoomedIn = true;
+    }
+    return zoomedIn;
+}
+
+
+const MapViewer = props => {
+    const { onCountryChange, onCityChange, countryChosen } = props;
+    const viewBoxInitValue = "0 0 1000 684";
+    const [previousTargetID, setPreviousTargetID] = useState('');
+    let highlightStyle =
+        "fill:#29B6F6;stroke:#black;stroke-width:0.11153841;stroke-miterlimit:4;stroke-dasharray:none";
+    let standardStyle =
+        "fill:#c0c0c0;stroke:#black;stroke-width:0.40000001;stroke-miterlimit:4;stroke-dasharray:none";
+
+    function handleMapClick(event) {
+
+        let targetID = event.target.id;
+        let previousTargetElement = document.getElementById(previousTargetID);
+
+        if (countryChosen !== '') {
+            previousTargetElement.style = standardStyle;
+            setPreviousTargetID('');
+        } else {
+            if (targetID !== 'svg2') {
+                event.target.style = highlightStyle;
+                setPreviousTargetID(targetID);
+            }
+        }
+
+        let zoomedin = zoomAndPan(document.getElementById('svg2'), event.target, viewBoxInitValue);
+        if (zoomedin && targetID !== 'svg2') {
+            onCountryChange(event.target.getAttribute('name'));
+        } else {
+            onCountryChange('');
+        }
+    }
+
+    return (
+        <svg
+            onClick={handleMapClick}
+            id="svg2"
+            baseProfile="tiny"
+
+            fill="#c0c0c0"
+            stroke="black"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth="0.40000001"
+            strokeMiterlimit="4"
+            strokeDasharray="none"
+
+            version="1.2"
+
+            viewBox="0 0 1000 684"
+
+            height="684"
+            width="1000"
+
+            xmlns="http://www.w3.org/2000/svg">
+
+            <path d="M654.7 528.1l0.5 0.4 2 2.9 1.4 0.5 1.9 1.3 1.4 3.2 0.1 2.2-0.5 2.6 0.3 2.1-0.8 0.8 0.7 2 0.2 1.9 1.2 2.2 1.2 1.1 1.3 2.4 1.6-0.2 1.3 1.1 0 1.1 1.1 1.8-0.8 2.6-1.7 0.8-1.2 3.1-0.3 2-0.6 0.5-1.9 0.3-1.7 1.3 1 2.2-0.9 0.7-0.3 1.5-0.7 0.7-2.7-0.9-0.7-2.5-1.7-2.7-4.9-2.6-1.2-1.1 0.4-1.5-0.1-1.4-1.4-2.4 0.3-2.6 0.8-2.2-0.3-2.7 0.1-2.1-0.7-2.9 0.5-2.1 0.9-1.3-0.2-2.2-1.5-1.1-1.6-0.2 0-3.1-0.3-0.6 1.7 0-1.7-2.8 3.2-5.3 1.1 0.3 0.8 2.1 3.4-1.2z"
+                id="AL" name="Albania">
+            </path>
+
+            <path d="M423.4 528.8l-2.8 1.1-0.7-0.4 0-2.1 0.9-0.7 2.6 0.6 0 1.5z"
+                id="AD" name="Andorra">
+            </path>
+
+            <path d="M974.6 572.4l-3.2-2.8-2.3-1.7-1.6-0.7-3.5 0.4-2.9-0.9-1.8-0.9 0.5-0.6-1.9-3-0.8-1.9 1-1.4 0.2-2.3-1.7-3.3-1.7-1.2-1.3-1.6 4.4-0.3 5.3-1.3 3.2 0.3 1.1-0.2 3.5-0.1-0.5-0.7 2.3-0.5 1.3 1.5 1.6 1-0.9 0.9 2.3 1 1.2 0 3.3 2.5 0 0.7-1.9 1.5-0.2 1.1 3.2 3.4 2.4 1.4 3.2 1.5 0.2 0.9-0.9 2.5-3 0 0.6 0.9 1.9 1.2 6.1 4.4 1.5-0.3 2.2 0.9 0.2 1.3-1 1.4 1.9 1.4 0.8 1.5-1.8 0.5 1.4 1.5 0.5 2.6-2.2-0.1-2.5 0.5-3-4.7 0.1-1-3.1-2.1 0.1-1.7-0.8-1.2-1.7 0.2-1.9 0.9-2.9-1.6-1.2 0.2-0.3-1.1-1.8-1.7-3.2 0.9z"
+                id="AM" name="Armenia">
+            </path>
+
+            <path d="M607.8 434.9l-0.9 2.4 0.1 0.8 1.6 2.9 2.3 2.9-0.6 0.7-0.6 1.9 0.7 2-3.9 0.3-2.5-0.9-1.3 1.4 2.6 0.7 0.5 1.1-0.5 1.4-2.2 1.2 0.4 1.5-0.5 0.7 1.1 1.6-0.3 1.8-2.5 0.5-1.8 1.7-1.3 0.9-0.1 1.9-2.4-0.5-1.6 0.2-2.3 1.1-2.8-0.2-2.6 0.2-1.9 0.7-1 1.3-3 1.2-5-0.7-5-0.9-2.6-0.5-4-0.3-8.6-1.5-1.1-0.5-3.1-3.5 0-2.3-4.7 1.5-3.1-0.2-3.4 0.3-1.4 0.6-1.2 2.2-1.1 0.4-2.1-0.4-0.9-0.8-3-0.3-0.5-1.5-0.8-0.3-2 1.9-2.3-0.4-1.5-0.8-0.4-1-3.3-0.8 0.3-0.8-1.1-2.5 1-3-0.9-0.7 1.7-0.5 0.7-0.5 3.5 1.6 0.9 1.4 1.2 0.3-0.1 1.3 1.6-0.5 1-1.6 0.2-1.9 2.8-0.1 2.7 0.4 1.4 1.9 3.8-0.5 1.1-0.9 3.9-1.5 5.7-0.5 0.1-1.4 2 0.3 1.5 0.8 2.5-0.5 1.1 0.5 0.2 1.1 1.2 0.9 1.7 0.4 0.3-1.5-0.3-1.9-1.7-0.3 0.6-1.3-0.1-1.3-2.5-2.8 0.6-1.3 3.1-1.8 2.9-0.8 1-1 0.7-2.9 2.3 0.9 1.3-0.9 0.2-2.8 2.2 1.2 0.8 1.3 3.9 0.4 1.4-0.7 2.5 0.4 0.1-1.1 1.3-1.6 1.2 0 0.3-3.2 1.2-0.2 1.6 0.8 1.4-0.4 4.8 1.7 1.5-0.1 3 1.7 3.8 0.2 1.2-0.9 5.1 1.5 1 1.6z"
+                id="AT" name="Austria">
+            </path>
+
+            <path d="M470 401.8l1.6-0.5 0.8 0.8 1.2 0.1 3 2.3-0.4 1.5 1.7 0.7 0.4 2-2.3 1.3-0.6 1.7-1.7-0.7-1.3 1.3-1.4 2.5-0.2 1.7 2 2.4-1.1 1.7-3.4 0.4-2.8-2.6-3.3-1.5-1.7-0.1-0.4-2-0.6-0.6 0.8-2.7-1.9 0.6-0.5 1.4-1.3 0.6-2.2 0.3-2.3-0.3-0.5-0.6 0.7-1.4-0.7-0.7 0.4-1.6-1.6-1.1-3.6-0.4-0.9 0.3-0.6-2-4.2-1.1-0.5-2-1.6-1.8-3.2 1-2.9-2.4 0-1.2-0.9-2.2 5.1-2.5 4.7-1.7 0.3 1.3 1.2 0.7 1.2-0.6 3.9 1.1 3.2-1.5 0.6-1.2 2.1 0.3-0.2-0.9 1.4-0.7 1.5 0.8 1.5-1 1.1 1.3 2.5-0.8 0.5 1.6 2.6 1.4 1.9-0.4 1.7 1.3 2.6 1.1-1 2.9-1.1 1.3 0.7 1.1z"
+                id="BE" name="Belgium">
+            </path>
+
+            <path d="M759.6 509.6l0.2 3.7-0.9 1.8-2-0.6-2.2 0.5-1 2-1.2 1.2 0 2.6 0.4 4.2-1.6 0.7-2.9 3.7 3 1.8 1.7 2.2 2.4 2.5 0.5 1.2-1.7-0.2-0.9 0.4-1.8 0-1.5 0.6-3.4-2.5-3.5 0.3-1.5 1.1-3.4 0.4-1 2.2-1.9 0.3-0.3 1.4-1.6-0.5-1.6 1.1 0.7 1 0.9 2.7-0.2 0.8-2.5 1.2-1.8-0.3-3.2 0.4-3.4 0.9-1.7-1.1-3.7-1.2-0.9 0.6-2.4-1.4-1.2-1.7-1.6 0.5-4.2-0.1-0.2 0.9-1.9 0.2-3 1.1-2.6-0.2-2.4 0.2-1 1-3-0.2 0-4.2 0.5-1.7-0.9-1-1.7-3.3-3.4-1.8-3.3-3.2 1-0.3 1.1-1.7-1.4-2.9 0-3.3 1.1-0.6 1.8-0.1 1.1-1.6 1.6-1.7 0.3-1.4-1.6-1-1.3-1.6-2.9-1.5-2.4-3.9-0.6-2.4 0.4-2.3 2.1-1 0.2-1.9 0.9-0.7 4.3 2.5-1.8 2 0.9 1.8 3.7-0.6 3.9 0.3 5.3 1 3.5 0.4 2.5-0.5 9.4 1.7 4.2 0.2 2.3-0.6 1.6-0.9 1.2-1.6 3.3-2.1 3.2-1.2 4.4-0.9 2.9-0.4 4.5 2.3 3.2 0.4 0.9 0.6 1.7-0.5 2.5 2.6 2.2 0.7 2.6 0.5 2 0z"
+                id="BG" name="Bulgaria">
+            </path>
+
+            <path d="M638.2 492.2l2.8-0.7 1.5 0.8-0.5 2.5-1.6 2.7-0.1 2.6 0.5 0.8 2.4 1.2 3.3 2.9-0.3 0.8-3.1-0.1-0.5 0.8 3.2 3.7-0.2 2.2-2.5-0.3-0.7 0.8-3.2 0.5 1.4 2.2-2.3 0.2-2 1.7-0.5 1.7 0.1 1.5-1.7 0.2-0.5 0.7 0.6 3 1.1 1.6-0.1 0.8-1.1 1-1.7-0.7-6.2-3.8-0.6-0.9-1.6 0.1-1.1-0.6 0.3-1.6-3-2.5-1.5-1.6-0.4-2.2-2.4-1.1-5-4.1-1.7-2-2.8-2.3-2.3-2.4-1.3-4-3.4-3.2-1.9-1.3-0.1-1.4 0.3-5 1.2-0.6 1.8 0.4 1.7 1.8 1.8 1 2.7-3.2 3.2 0.3 1.5-1.2 2.7 1.6 2.5 0.1 2.8 0.7 1.7-0.6 1.6 1.3 2.2-1 1.8 0.3 1.8-0.2 2.6 0.7 2.2 0.2 1.1 0.8 1.3 2.2 2.2 0.2z"
+                id="BA" name="Bosnia and Herzegovina">
+            </path>
+
+            <path d="M724.5 321.5l1.8 1.3 1.2-0.4 1.9-0.1 3.1 2.2 1.6-0.9 1.5-0.3 3.6 1.2 0.2 2.3 1.7 1.5 2.1-1.3 2.7-1.1 3.4 0 2.7 0.9 2.4 1.8 3.5 1.4 0.4 2.5-0.6 1.8 2.1 2.1 0.5 1.3-1.3 2 0 2 4.1 2 0.5 0.5-0.5 1.8 1.4 0.6 1.8 2.4 1.5 1.4 4.9 2.1 0.7 0.8-0.4 2.8 5.3 0.5 3.2 1.3 0 1.1 0.7 1.1 2.8 1.4 0.6 1.7-2.6 1-0.3 0.9-3 1.8-3.5-0.2-2.4-1.5-2.9 0.1-1.4 2.7 3.9 3.3 0.6 1-0.2 1.5 1 1.5 0.5 3.3 1.6 1.4 1.3 1.8-5-0.1-1.4 0.8-1.7-0.4-1.1 0.5-2.2 2.3-0.8 1.2-1 3.3 1.8 3.6-0.8 1.4-2.7-0.9-0.5-1.2-2-1.1-7.1 0.6-2.3 0.8-0.6-0.5-2.5-3-3.8 1.6-0.6 1.1-1-0.3-0.8-1.3-5.1-1-1.9 0.7-1.9-0.5-1.2 1.7-0.8-1.4-1.1-0.4-3 0.1-1-0.3-1.8-2.1-2.3 0-6.1-0.9-6.5-1.5-7.8-0.4-2.4 0.1-2.3 0.5-5.9 0.5-0.8 1.4-1.6 1.6-1.9 1.2-1.1-0.5-1.9-0.3-1.1 0.5 0.1 1.4-1-1.6-0.1-1.3 0.8-1.5-0.4-1.1 0.2-2.4-0.8-1-1.8-1-3.3-1.3 0-1.2 2-2.7 0.8-0.5 4-1.7 0.6-1.6-0.9-3.9-1.7-3.6-2.9-4.9-2.1-5.1 2.9 0.4 1.6-0.3 3.7-0.3 1.6 0.9 3.3-1.3 1.8 0.1 0.8-2.6 2 0.2 1.4-1.2 2.1-0.5 1-0.7 1.7 2.3 2-0.2 0-1-0.9-1.1-1.9-0.5 0.5-2 1-1.5-0.4-2.3 1.1-3 2.6-0.6 0.8-0.6 1-2 3.6 0.1 0.4-1.1 1.3-1-3.9-1 0.7-3.1 0-1.8 2.6-0.6 1.2-1.5 1.1-0.3 3 0.4 3.1 0.1 0.4-1.7 2.5-2.4 1.4-0.8 1.3-0.2z"
+                id="BY" name="Belarus">
+            </path>
+
+            <path d="M514.1 449.1l0.8 1.1 2.6 1.3 1.2 0 0.9 0.7-1 3-0.5 1.5 0.1 1.7 1.2 0.1 3.3 0.8 0.4 1 1.5 0.8 2.3 0.4 2-1.9 0.8 0.3 0.5 1.5-0.5 3.1 0.5 1.8-1.9-0.2-1-0.9-1.3 0.3-0.6 1.8 1.5 3.5-1.2 0.3-1.4-1.9-0.7-0.1-2.9 1.1-2-0.8-0.8-2.1-2 0.1 0 2.9-0.5 1.1-2.2 2.6-0.2 1 0.7 1.7-1.2 0.7-0.9-1.4-1.3-1.2 0.4-1.2-2.2-0.6-2.4-2-0.2-2.5-1-0.6-3.4 2.9 0.6 1.5-1.6 2.2-2.4 1.4-3.1-0.8-2.6 1-2.4 0.5-1.4-0.7-0.9-1.4-2.3-2.3 0.5-1.7-0.8-2.2-2.2-0.3-1.9 0.1-2.3 1.5 0.5 1.3-2.3 1.6-1.4-0.1 0-1 1.5-1 0.3-1.5-0.8-0.7 1.1-2.9 3-2.2 0.5-2.9 2.1-0.9 3.8-4 0.6-0.9-1.3-1.1 1.8-1.4 1.1 0 0.8 0.8 2.7-0.3 0.8-1.4 1.5-0.7 1 0.3 2.8 0.1 3.3-0.6 2.6 0.2-0.2-1.5 1.3-1.2 1.3 0 1.4 1 0.8-0.2 1.1 0.9 3.7-0.2z"
+                id="CH" name="Switzerland">
+            </path>
+
+            <path d="M578.9 400.5l2.1-0.4-0.1-1.9 3.3 0.8 0.3 1.1 1 1.1 4.6 1.1 2.7 1.1 0.8 0.9 3.2-0.7 1.8 1.2-0.7 1.1-1.6 1.2 1.6 0.9 2 1.8 1.5 1.9 1 0.4 3.5-2.5-1.7-2.4 1.3 0 2.2 0.6 3.3 1.9 3.6-0.7 0 1.6-1.4 0.5 2.7 2.3 1 0.6 1.8-1 3.5 1.8 0.5-0.2 2.7 0.8 0.5 1.8 2.8 2.2 0.5 1.6-2.8 0.3-2 1.5-0.6 0.8-2.3 1.2-0.7 3-1.6 0.8-2.1 1.9-1.5 0.7-1.7 0.2-3.6-0.5-1.4 1.2-1.1 2.8-1-1.6-5.1-1.5-1.2 0.9-3.8-0.2-3-1.7-1.5 0.1-4.8-1.7-1.4 0.4-1.6-0.8-1.2 0.2-0.3 3.2-1.2 0-1.3 1.6-0.1 1.1-2.5-0.4-1.4 0.7-3.9-0.4-0.8-1.3-2.2-1.2-0.6-0.7-2.9-2.2-1.3 0-1.3-1.6-2.5-1.5-2.9-2.6-1.3 0-2.3-2-1.8-2.7-1.4-1.5 1.2-1.7-0.6-1.6-2.3-1.3-2.5-3.4 0.5-0.7 1.9 2 1.9-2.6 1.1-0.6 2.6-0.6 2.1 0.4 0.8-1.3 2-0.3 0.6-1 1.6-0.7 1.1 0.1 0.6-1.2 4.4-1 3.5-1.5 1.9-0.6-1.5-1.5 0.8-0.6 2.2 0.4 1.5 1.5 0 0.9 1.4 0.6 0.9-0.7z"
+                id="CZ" name="Czech Republic">
+            </path>
+
+            <path d="M567.9 355.2l-3.4-0.1-0.2-1.7-1.1-1-0.1-1 4.7 2.6 0.1 1.2z m-34.8-8.1l-1.8 0-1.4-0.7 0.8-1 1.7 0.4 0.7 1.3z m28.3 0.5l-1.3 0.7-1.3 0-1.3 1.3-2.1-1.1-0.5-1.1 0.1-2.2 0.6-1.4 2-0.9 0.9 1.2 1.7 0.6-0.5 1.7 1.7 1.2z m-46.6-6.6l1.8 0.7 1.6 1.6 0.1 1.4-1.7 1.6 3.2-0.3 0.8 1.2 1.7-0.4 4.5 1.8 3.2-0.9 0.7 1.4-0.6 1.5-2.2 1.6 1.4 1.2 2.1-0.2 3.6 0.9 3.3-2.5 1.1-0.5 3.6-0.3 5.1-4.5 5.3 0.9 1.5 1.9 3.7 2.1 3.2-0.2 1.3 2 0.8 2.5 1.9 1.3 2.8 0.5 0.7 2.6 1.8 4.1-0.3 2.7-1.9 1.8-0.7 1.6 1.6 1.4 3.3 2.1 1.4 1.7-0.6 2.6 2 2.3 0.8 1.9-0.7 2.6-0.8 1.1 1.9 3.1 0 1.6 2.2 0.9 1.6 3.2-0.4 2.3-1.6 3.6-0.9 0.7-1.4-0.6 0-0.9-1.5-1.5-2.2-0.4-0.8 0.6 1.5 1.5-1.9 0.6-3.5 1.5-4.4 1-0.6 1.2-1.1-0.1-1.6 0.7-0.6 1-2 0.3-0.8 1.3-2.1-0.4-2.6 0.6-1.1 0.6-1.9 2.6-1.9-2-0.5 0.7 2.5 3.4 2.3 1.3 0.6 1.6-1.2 1.7 1.4 1.5 1.8 2.7 2.3 2 1.3 0 2.9 2.6 2.5 1.5 1.3 1.6 1.3 0 2.9 2.2 0.6 0.7-0.2 2.8-1.3 0.9-2.3-0.9-0.7 2.9-1 1-2.9 0.8-3.1 1.8-0.6 1.3 2.5 2.8 0.1 1.3-0.6 1.3 1.7 0.3 0.3 1.9-0.3 1.5-1.7-0.4-1.2-0.9-0.2-1.1-1.1-0.5-2.5 0.5-1.5-0.8-2-0.3-0.1 1.4-5.7 0.5-3.9 1.5-1.1 0.9-3.8 0.5-1.4-1.9-2.7-0.4-2.8 0.1-0.2 1.9-1 1.6-1.6 0.5 0.1-1.3-1.2-0.3-0.9-1.4-3.5-1.6-0.7 0.5-2.2-1.2-5.1-2.3 1 1.6-3.7 0.2-1.1-0.9-0.8 0.2-1.4-1-1.3 0-1.3 1.2 0.2 1.5-2.6-0.2-3.3 0.6-2.8-0.1-1-0.3-1.1-1.3 0-1.5 0.6-2-0.2-2.4 0.3-1.5 1-1.8 1.4-5.5 3.5-3.8-0.1-1.3-1.6-0.6-4.8-0.8-2-1.4-3 0.6-2.1-0.1-0.3-0.9-1.4-0.4-1.8 0.7-3.5-4.3-1.4-0.1 0.6-3 0.9-0.9 0-1.4-2.8-1.1-1.5-1.6-0.3-2.2 0.6-1.7 2.3-1.3-0.4-2-1.7-0.7 0.4-1.5-3-2.3 0.6-2.4-1.9-1.2 0.5-0.8 2.2-1.6-0.7-1.2 1.3-2.8 0-1.2-3.1-4.1 0.7-1.1 1.8-0.7 2.3 0.8 4.5-1.3 0.7-1-1-0.9 0.3-0.9 2.6-1.6 0.6-2.6-0.8-1-2.6-0.3-0.7-1 0.6-1.6 3.1 0.1 1.1-3.8 0.7-1.7-0.1-4.3-1.7-1.4 0.5-2.7 1.1-1.4 0.9-0.4 4-0.3 4.5 0.1 1.9 2.2-0.6 1.2 1.6 0.3 0.9-2.4 1.4 0.8 1.1-0.1-0.5-1.7 0.2-1.6 0.9-1.4 3.3 0.6 3.6-0.3 0.1-0.6-2.8-0.5-0.9-1.1-0.2-3.9-1.5-0.8-1.6 0.3 0-1.5 3.4-1.1 0-1-3.4-3.8-0.2-1.6 2.7 0.1 5.2 1.3 3.1-0.7 1.5 0.4z"
+                id="DE" name="Germany">
+            </path>
+
+            <path d="M533.5 340l2.1 1 2.4 0.3 0.4 1.9-3.5 0.7-5.1-2.1 0.1-2.5 2.3-0.2 1.3 0.9z m13.7-1.1l-2.3 0.1-1.9 1-0.3-1.2 1.6-1 1.2 0.2 1.7 0.9z m-28.8 1.2l-1.2 0.2-1.7-0.5-0.6-2.3 2.7 1.1 0.8 1.5z m8 2l-1.4-1.5 2.5-2.9 0.7-0.2-1.8 4.6z m50.1-4l-2.4-0.2-2.4-1 0.1-2 0.5-0.9 4.5 2.3-0.3 1.8z m-52.1-8.7l2.3 4.3-0.1 2.8-1.8 1.2-2.1 0-5.4-1.7-1.7-2.9-0.1-2.3 4.8-1.4 2.6 0.8 1.5-0.8z m23.3 0.2l-1 0.6-0.7-0.9 0.9-0.9 0.8 1.2z m-23.9-2.8l-0.9 0 0-3 1.5 1.6-0.6 1.4z m22.6 0l0.1 1.5-0.7 0.7-2 0.7-1.1 1.8 2 1.2 0.5 1.5-1 0.7-2.6 0.7 0 5.6-2.2 0.6-1.5-2.1-0.1-0.9-1.3-3.1-2-0.4-2.2 0.1-1.5-1.8 0.1-2.1-1-2-1.3-0.6 0.4-1.4 3.1-0.2 1.5-1.9 0-0.9 1.8-0.2 0.8 0.7 0.1 2.6 1.6 0.5 1.1-2.9-0.9-1.1 3.9-2.2 1.1-0.1 3 0.9-0.4 2.1 0.7 2z m-31.6 14.2l-1.5-0.4-3.1 0.7-5.2-1.3-2.7-0.1-0.5-2.1-0.8-1.4 1.1-0.3-0.9-3.9-3.2-1.3-2.5-1.4 0.5-5.6-1-2.3 0.1-6.9 1.3-0.1 3.1 0.8 1.5 0.9 0.6-1.9 1.6-1.6 2-0.8 1.6 1.3 0.3-4.5-1.6-0.5-1.3 0.4-1.3 1.9-1 2.4-1.9 0.2-1.5 0.7-2.4-1.4 0.1-1.4 1.6-1.9 2-1.9 2.2 0 1.6-0.6 3.9 0.1 2.8-1.3 2.8-3.6 1.5-1.5 3.3-0.5 3-1.7-0.8 2.5 1.2 1.7 0 3-0.9 1-1.4 2.5 0.2 5.6 2.5 1.5 4.1 0 1 1.1-0.7 2.2-2.6 1.4-0.9 0-1.4-1.1-1.1 0.9-1.2 5.1-2-0.1-1.2 0.5 1.4 1.2-1.3 0.8-3.4 3.1 1.1 3.4-0.2 0.9-1.5 1.3-0.6 1.1 2.3 0.3 1.3 2.9z"
+                id="DK" name="Denmark">
+            </path>
+
+            <path d="M655.5 285.3l3.9 0.2 4 1.8-1.8 0.7-3.7 2.5-0.3 0.5-2.6-0.1-1.4 0.2-1 0.9-0.9 2.8-1.6 0.5-0.3-0.9 1.6-1.8-1.5-1.4-1.9-0.7 0.9-1.8-1.6-1.6 1.5-0.2 1 0.5 2-0.5 0.5-1 3.2-0.6z m8.3 1l-0.9 0.2-2.4-1.1 1-1 1.9 0.5 0.4 1.4z m-5.4-4l-0.8 0.7-0.9-0.6-0.9 1.6-1.3 0.3-1.9-2.5-2.9-0.5-1.3-0.7 4.5-0.4 1.2-1.6 1.3 0.1 0.4 0.9 2.1 0.3 0.9 1-0.4 1.4z m53.2 18.9l-1.9-0.2-3.4-1-4 1.2-2.8-1.1-3.5-2.7-0.6-0.7-2.3-0.5-5.6-2.5-1.1 0.7-1.5-1-8.4 2.8 1-3.4 0.1-3.6-2.3-0.4-0.9 1.3-1.3 0.3-4.1-1.3-1.2-2.1-1.6-1.1-0.8-1.2 0-1.8-1.5-2 0.7-1.1-0.7-2.8 3-1.2 3.3-0.2 0.9-1.3 2-1.3 2.3 0.2 3.1-1 6.3 0.1 0.4-1.7 3.2 0 7.6 1.2 1.8 0 2.8 1.2 1.4 0.3 4 0 6.3 0.5 1.2-1 1.5 1.2-1 1.4-1.2 0.4-0.9 3.2-0.8 1-1.7-0.3-2.5 0.2-2 0.5-1.3 1.1 0.1 1.1 2.6 2.3 1.4 2.8 2.6 2.2 0.8 1.3 2.4 2.9 1.8 2.5-2.3 0.6-1.2 2-0.2 2z"
+                id="EE" name="Estonia">
+            </path>
+
+            <path d="M646.5 260l2 0.2-0.6 2.8-2.3-1.8 0.9-1.2z m-1.9 0.5l-2.1-0.6 0.9-0.8 1.2 1.4z m-6.5-2.7l-1.5 0.7-0.3-1.7 1.8 1z m25.5-62.9l-1.5 0.5-1.3-0.3-0.2-0.9 2-0.6 1 1.3z m30.1-54.6l-3.7 1.3 1.6 0.8 0.8 1.2-2 3.2 0.1 0.6 3.1 4 7.2 1.7 2.3 1.8 3.6 2.3 1.8 0.9 0.1 1.9-3 3-2.6 3.1-1.2 1.7 0.4 1.7 2.6 2.1 2.3 2.2 3 3.5 3 2.5 3 4.2 0.4 1.4-3.7 0.7 1.2 0.9-0.6 1.3 0.3 1.9-0.5 1.5 1.9 0.3 0.3 1.3-1.7 1 0 0.9 1.4 1.7 1 0.6 3.2 0.6 0.9 1.6-1 1.7 1.8 2.2 4.5 1.8 0.7 1.3 0.2 1.7-2.4 3-2 0.8 0.7 0.8 4.5 2.6 6.6 2.9 2.6 1.3 2.1 1.9 2.2 1.6 0.5 1-1.6 4.5-0.7 1.2-2.2 2.2-3.4 2.8-2.3 2.3-5.3 5.9-1.9 1.5-2.1 2.2-4.9 3.3-0.7 0.8-2.5 1.5-5.6 5.2-3.5 1-2.6-1.1-1.8 0.2-1.2 0.8-2.5 0.2-2 0.6-1-0.7-0.7 0.6-1.9 0.2-2-1 0.3 1.5-3.3 1.5-0.8-0.9-2 1-2.1 0.2-1.1 0.8-5.9 1.1-1.5 1.3-1.2-0.3-6.7 1.1-2.9-0.2-2.7 2-1.8 0.4 1.5-3-2.4-1.1-1.2-1.6-0.7 0.4-0.3 1.5-1.5 0.8-2-0.1-0.4-2.4 0.8-2.3-3.6-0.6-3.8-1.4-0.9-0.1-0.7-1.3-2 0.9-2.1-0.9-1.7-5.3 0-1.3 0.9-1.6-0.1-3.4-0.6-2 1.1-0.6-0.9-1.6-2.2-3-1.6-1.1 0.5-3.3-0.6-1.6-1.9-1.1-1.1-2.9 0.6-2.4 2.4-2.2 0-1.2 0.9-1.1-0.5-1.3 2.7-0.5 1 0.4 2.4-0.4 1.9-1-1.1-1.8 1.5-0.7 1.2-1.2-0.1-1 2.3-0.5 2.4-1.9 2.3-1.1 2.3-1.9 1.1-0.1 0.3-1.3 2.5-1.9 0.8-1.6 2.3-1.8 2.2-3.9 3.8-1.2 3.8 0.3-0.6-1.7 1-0.3-0.6-1.3-1-0.6 0-4.6-1.3-0.9-4.5-1.6-1.7-0.1-1.2-1.2-4.6-0.7-1.6-2.5-1.9-2.2-1.8-0.6-1.3-2.6 0-1.3 1.5-1.3-0.1-1.8 0.7-1.4-0.6-0.9-3.3-2.6-0.8-1.6 0.1-1.1 1-0.8-0.6-1.3-2.8-0.4 0.3-1.9-0.7-1.4-0.4-2.4 1.1-1-1.8-0.9-1.9-1.5-1.4-0.1-1.2-1.6-3.5-1.8-8.3-1.7-4.2-1.8-4.8-2.3-2.9-1.1-0.4-0.9-2.9-0.8 0.5-0.4 2.1 0 1.9 0.4 0.5-0.6-1-1.4 1.9-1.1 3.6 0 7.9 5.5 1.4 1.8 6.9 0.6 0.8 0.5 1.8-0.1 3.7-0.9 1.2-1.2 1.5 0.1 3.6 1.1 3.9 0.8 2.6 0.8 1.3-0.6 0.5-1.7 1.6-1.3 2.1-0.3 1.5-1.4-0.7-3.2 0.6-1.7 0.7-3.8 2.6-1.7 1.7-1.8 2.1-0.2 3.8 0.4 0.9-0.2 3.5-1.6 2.9-0.3 1.8 1.2 4.2 2 4.2 1.1 3.6 0.8 2.6 2.7-0.7 1-3.2 2.9 0 0.7 1.7 1.3z"
+                id="FI" name="Finland">
+            </path>
+
+            <path d="M389.1 403.1l-1.6 1.3-1.3 0.2-2.5-1.2 0-0.5 2.5-1.1 2 0.6 0.9 0.7z m-36.5-39.7l1.3 0.9-3.4 1.9-1.2-0.6-0.9-1.2-0.1-1.9 1.2-0.5 1.8 0 1.3 1.4z m-23.3-11.4l-2.1 0.4-2.9 0 0.1-1.5-1.4-0.5-1.8-2.7-1.3-0.5-1.9 1.6 0.5 0.9-2.4 1.8-2.9-0.3-0.8-0.7-2.1-0.4-0.2-1-2.7-2 1-1.4 2.9-0.9-1-1.4 0.2-0.8 1.6 0.1 2.2-0.8 1.2-1.6 1-2.3 1.9-0.9 1.3 0.6 1.9-2 2.8-0.1 2.7-0.7 3.9 0.3 1.1 1.1 0.5 1.7 1.3 1.7 1.7 1.4 0 0.9-2 1.1 3.5 0.3 1.2 2.5-2.1 0.3-0.1 1.4 1.1 0.2-0.6 1.4-2.6 0.6-1.4 2.2-1 0.5-2.3-0.5z m13.7-20.2l-2.6-0.5-0.6-2 0.9-1.4 1.8 0.7 0.6 1.3-0.1 1.9z m-11.5-7.1l0.7 3.4-2.4 1.3-0.8-2.4-1.4 1 0.2-1.9 1.5-0.3 2.2-1.1z m1.7 1.7l-1.1-1.2 1.6-2.2 1.7-1-0.1 1.6-2.1 2.8z m2.5-7.8l-4.6 0.8 0.6-3-1.9-0.9 0.3-0.9 2.1-0.5 1.7 1.7 2.1 0.6-0.3 2.2z m-16.2-11.4l-1.5-0.2-0.4-1.4 0.2-2.3 1.6 0.2 0.1 3.7z m12.8-5.7l-0.1 2.8 2.9 0.8 2-0.1 0.4 0.6-3.3 2.7-0.7-0.1-0.2-2.2-1.4 0.3-1.9-0.3-1.2-1.9-2.7-0.5-0.9-1.1 1.4-0.6 0.3-1.2 2.8-0.6 1.5-0.7 0.9 1 0.2 1.1z m-11.9-2.6l1.2 0.9-1.1 1.3-1.5 0-2.2-1 0.2-0.5 3.4-0.7z m11.8-9.9l-1.6 2.6-1 0.7-0.2 1.7-3 1.6-3.2 2.4-1.4-1 2.6-1.7-2.1-1-0.2-2.6 0.8-0.7 1.4 0.7 1.9-0.3-0.7-1.2 6.3-3.2 0.4 2z m34.8-2.2l0 1.2-1.3 1.6-2.2 1.2-4.2 2.8-2.5 1.3-0.5 1.6 1.4 0.1 0.2 0.8-2.2 1.6-0.6 1.5 3-0.4 5.3-1.5 3.6 0.5 2.6-0.3 6.9 0.2 1.9-0.3 2.3 1.3 1 1.7-2.8 3.4-0.8 2.3-2 3.5-1.9 2-2 2.5-2.1 1.1-3.1 0.5 4.5 1.9-1.4 1.8-2.4 0.1-2.3 1.7-2.1 0.7 3.6 1.1 2.4-1 2.7-0.1 5.2 1.9 3.6 3.4 2 1.5 0.5 1.1 2.1 6.9 1.5 3.8 1.6 1.8 4.5 1.6 1.1 0.6 3.4 3.3 3.4 2.4-1.4 1.5 0.4 1.2 2.1 2.9 0.3 1.6-1.3-0.1-2.3-1.4-0.3 0.7 5 3.3 1.6 2 1 2.7-0.6 1.1-3 2.7 2.8 1.4 1.2-0.2 2-2.1 1.7-0.2 4.2 0.3 3.8 1 3.3 2.1 0.7 1.1 0.4 3.1-2.2 5.8-2.8 1.9-0.5 1.7-2.1 0.9-1.7-0.3-2.4 1.2 1.7 0.6 0 1.7-1.1 0.6-2.5 0.2 1.1 2 2.5 0.4 4.3-0.2 1.9 0.2-0.2 2.7-0.4 0.4-3.8 1.7-1 1.8-2.2-0.1-1.1 0.7-5.7 1.9-4.9-0.8-2.9 0.1-4 0.6-4.2-1.2-1.8-0.2-2.8 1.7-4.2 0.5-2-0.2 0.9 1.5-0.9 0.4-3.8-0.5-1 0.5-1.4-0.2-2.7-1.4-2.6-0.1-4.9 1.2-1 1.3-1.3 3.5-1.1 1.2-1.4 0.2-4.5-2.5-1.5 0.5-2.5 0.3-2.7 0.7-3.9 2.4-0.9 1.5-1.3 0.2-1.1-0.9-1.4-0.3-2.3 0.8-0.3-1.3 1-1 2.8-0.7 2.5-1.9 2.2-2.3 4.6-3.8 0.8-3.1 2.7-0.7 1.4-2.5 4.1-0.6 2.8 0 2.8 0.5 2.9-0.1 1.1-0.7 1.9-2.4 3.6-3-4.7 1-3.8 2.3-3.2-0.4-2.3-1.9-2.1-0.8-2.1 0.4-0.7-0.9-2.4-1.7-2.6 0-3.6 1.7-2.6-1.2-0.9-2.3 0.3-1.4 1.1-0.7 2.5-0.6 3.8-1.7 2.2-0.7 2.8-1.9 1.2-2.3-0.3-2 0.5-1.4-0.9-1.7 0.3-1.5-3 0.3-2.7 1.2-0.7-1.1 2.8-1.8 1-1.4 2.6-1.7 3.6-1.3 1.9 0.1 3.7-0.7 2.7 1.3-0.8-2.1 1.2-0.4 0.1-2.4 1.6-2.2-1.2-0.6 0.1-2 1.5-0.8 0.7-2.6-1.7-0.5-2.1 0.6-1.7-1.5-2.8-3.6-0.3-1.4 1.6-3.1 2.3-2-1.9-0.8-1.4 0.2-2.7 1.2-2 1.3-2.1 0-0.8-0.9-1.1 0.1-2.5 1.1-3.5-1.3-1.1 2.3-2.6-2.5-0.3-1.9 1.3 0 1.2-2.4 2.9-4 0.6-1.2-0.5-1.4-1.9-1.5 0.3-2.6 0.8-1 1.5-0.4-1.9-1.2-2.9 0.9-1.8 0.9 0.3-2-0.6-0.4-1.5 1.4 0.1 3.3-2.1 5.7-1.1 0.9-1.4-0.5 1.1-3.9 1.4-2.5-0.6-0.3 0.3-3.6 1-2.9 1.7-3.8 2.2-3.5-4.1 2.7-1 0.5-2.4-0.4-0.8-0.7-1-2.3 1.8-0.2 1.7-1.1-1.5-0.7 1.5-0.9 1.7-2.1 0.4-1.9-1-1.5-1.6-0.6-0.2-0.9 1.7-1.9-0.7-1.5 1.6-3.1 3 0.1 1.9-0.4-2.3-2 0.7-2.9 3.7-0.3-0.8-1.8 0.3-2 1.6-0.9 1.3 0.2 1 0.9 2.6-0.9 0.6 0.8 2.8-0.6 3.7-0.3 4.5-0.6 2.2-0.4 2.3 0.2-0.6 1.7z m-0.6-4l-1.3 0.1-1.3-1.4 1.5-0.3 1.1 1.6z m1.3-3.4l2.7 0.6-0.6 1.4-1.4-0.7-2 0.3-1.9-1 0.4-2.1 1.7-0.1 1.2 0.5-0.1 1.1z m3.7-2.3l-1 0.3-0.4-1.2-1.7-0.6 0.4-0.8 2.7 2.3z m16.1-19.5l1.6 1.8 1 0.5-1.3 3.8-0.2 1.6-1 1.7-0.8-0.4 0.9-2.9-0.5-1.1-1.7 0.3-0.4-0.9 1.6-1.4-0.8-2-1.1-0.2 1.5-1.6 1.2 0.8z m2.9 0.4l-1.4-1.3 1.3-0.6 0.1 1.9z"
+                id="GB" name="United Kingdom">
+            </path>
+
+            <path d="M972.9 547.7l-2.3 0.5 0.5 0.7-3.5 0.1-1.1 0.2-3.2-0.3-5.3 1.3-4.4 0.3-1.3-1.2-1.9-0.2-1.1-1.7-3.5-2.5-1.2-0.4-0.6-1.1-1.4-0.3-1.4 0.4-0.9 1.9-3.4-0.8-3.6-0.1-1.1 1-4.3-1.3 1.9-3 0.4-1.7-0.4-2.4-1.7-2.7-3.7-8-1.1-1.2-4-1.4-1.1-1.6-3.2-2.1-4.1-0.9-4.8-2.9-2.9-1.7 0.9-2.1 0.8-0.2 4.8 0.9 1.5-0.4 5.9 2.5 3.6 0.7 3.1 1.7 6 0.5 2.8-0.6 1.8 0 2 1.1 2.4-0.2 3.2 1.2 1.5 1.6 3.5 1.4 2.8 0.8 3.2 1.5 0.2 2.4 2.5 0.4 4.3-2.1 2.1-0.7 1.8 0.2 1.9 1.9 0.8-2.2 4.6 1.7 2.1 1.8 2.8-0.1 2.2 1-0.2 4.2 2.9 1.4 1.6 1.2 3.4 0.7 3.1 1.5-1.2 2.1-1.2 0.3 0 1.3 1.1 0.8 0.9 1.5 4.5 2.6 0.9 0.8-0.2 2-1.8 1.4-4.1-2-1.7 0.5-3.2-0.9-1.4-1 0.1-0.7-4-1.4-1.9-0.4-3 2.5z"
+                id="GE" name="Georgia">
+            </path>
+
+            <path d="M712.9 637.7l2.3 0 1.7-1 2.2 3.7 1.7-0.1 3.5-0.9 3.7 0.2 1.4 1 2.6 0.2 2.3 0.5 3.3-0.7 0.4 0.4 0.1 2.8 1.2 0.2 0.7-0.7 1.7-0.6 1.8 0 1.6-0.8-0.4 3.5-0.9 0.4-4.4-0.1-8.1 1-5.3 0.4-0.7-0.2-0.7-2.2-3.3-1.1-6.2-1-1.5-0.4-3.1 0.2-1.1-1-0.4-1.8 0.2-1.9 1.4 0.3 0.4-1.4 1.9 1.1z m43.4 1.1l-1.6-2.1 1.1-0.5 1.2 2.4-0.7 0.2z m-54.7-11.4l-1.4 0.2-0.7-2.2 0.5-1 2.1 2.1-0.5 0.9z m62.7 4.1l-1.7-0.4 0.4-1.8-0.7-1.6 2.3-2.7 3.2-1.3 0.9 0.9-0.9 2.5-0.9 1.3 0.4 1-1.6 0.2-1.4 1.9z m-19-10.3l-1.5 1.2-0.9-0.6 1.9-1.2 0.5 0.6z m6.2-2.2l0-0.7 1.3-1.1 2.5-1 1.2 0.6-2.5 1.4-1.5 0.1-1 0.7z m-22.2-5.4l-0.9 1.3-1.3-0.7 0.4-1.2 1.1-0.6 0.7 1.2z m3.6 1.6l-1.1 0.6-1.4-2.2 1.9-2 0.9 0.7 0.1 1.8-0.4 1.1z m5.4-8.8l0.6-1.6 1.6-0.1 0 1.1-2.2 0.6z m-21.7-0.7l-1-0.4 0.4-1.2 1.1 0.5-0.5 1.1z m31.6-3.7l2.8 0.7-0.7 1-1.6 0.9-1.7-0.9-1.2-1.2 2.4-0.5z m-76.7 0.1l0.4 1.2-1.1 1-1.6-1.2-1.2-1.8 0.9-1.1 0.9 1.2 1.7 0.7z m53.1 0.7l-1.4-0.1-2.4-2.1 0.9-1.4 0.9 1.2 1.3 0.1 0.7 2.3z m-57.3-9.7l0.3 1.8 0.9 0.3 1.3 1.6-0.2 1.2-2.1-0.8-1 0.2-1.1-2-1.3 0.2 0.6-1.8 1.4 0.1 1.2-0.8z m70.8 2.5l-1.1 0.9-1.8-1.7 1.4-1.3-2.1-2.4-0.1-1 2.1-0.5 1.3 0.9 0.6 1.2 0.4 2.6-0.7 1.3z m-70.2-6.1l-1.3 0.1-0.4-0.9 0.8-2.4 1.1 0.3 0.2 2.5-0.4 0.4z m34.6-5.4l0.9 1.7 5.5 2.5 2.8 0.3 1.7 2.6-0.2 0.6 1.5 3.8 1.1 0.9 2.7 0.3 0.4 1.9-0.9 0.8-1.9-0.7-2.1-1.6-0.4-1.3-2.2-2.9-3.6-0.2-1.5-0.7-0.6-1.7-1.5-0.9-2-1.9-1.5-1.1-3-1.1-0.6-0.7 1.9-1.3 2.1-0.5 1.4 1.2z m37.9-5.8l-0.1 0.9 1.9 1.5 0.7 1-1.2 2.1-4-0.8-0.8-0.8 1.6-1.3-0.9-0.5-1.2 1.5-2.2-0.7-0.9-0.9 0.7-1.4 1.5 0.1 1.7-0.8 0-0.6 2.3-0.2 0.9 0.9z m-81.3-1.6l-1.3 0.3-1.2-0.8-1.2-1.9-2.2-2.2 0.7-1.1 1.6-0.4 1.4 1.2-1.1 1.2 0.8 0.7 0.9 2.4 1.6 0.6z m67.8-8.6l-0.7 1.4-0.1 1.3-3.8-0.7-0.3-2.3 2.2-0.1 0.8 0.8 1-0.9 0.9 0.5z m2.4-6.9l-1.4 0.4-1.7-1.3 1.4-0.5 1.7 1.4z m-11.9-3l-1.6 0.6-1.8-1.1 1.1-2.3 1.3 0.1 1 1.3 0 1.4z m15.9-1.7l-0.5-0.7-2.1-1.2-4.6-0.6-2.3-0.9-0.9 0.2-2-1-1.2 0.4-2.5 1.7-1.5-0.2-1.6-1-2.2 0.4-1.7 1.9-1.8 1-1.8-0.4-2.3 0-0.2 1.1 2 2.1-0.5 1 0.5 1 2.1 0.1 2.5 1.3-0.7 0.7-3.2-1.2-1.2-0.1-1.2 1.2 3.5 2.7 0.5 1.4-0.6 0.9-1.5-0.8-2.5-3.2-3.1-0.6-0.4 0.6 1.2 2.4 2.9 1.7-0.6 0.6-3-1-1.1-1.6-0.4-2-5.5-2.8-1.4-2.8-2.2 1.3 0.4 2-0.6 3.8 0.4 1.3 3.5 3.7 1.3 2.7 2.6 2.1 2.6 3.2 0.7 1.8-1.3 1.1-1.3-0.5 0.3-1.9-2.2-1.2-1.9 1.2 1.5 2.3 0.5 1.2-0.4 1.3-1.6 0.8-2.6 0.1 2.4 1.4 3.3 1 1.6 1.1 1.5 0.1 1.7 2.1 2.6 0.6 1.7 2.1 2 0.5 1.7 0.8 1 2.1 0.3 2.9 0.6 2.1 0.1 1.7-1.1 0.5-1.3-1.6-4.2-3.6-1-0.4-1 0.7-2.9 0.5-1.9 1.3 1.5 1.3 0.1 1.2 0.8 1.5 2 0.4 2 2.4-3 1-1.1 0.7-0.8-0.5-0.2-1.2-4.3-2.2-0.6 0.6 0.7 2.4 1.3 1.6 2.2 4.3 1.3 3.7-0.2 2.1 1 1.5 0.1 1.5-1.7-1.2-2.2-2.5-0.8-1.5-2.2 0.1-1.3 3.3 0.2 1.9-1.6-1.1-0.2-2.9-2.1-2.8-0.9-0.4-1.3-1.9-1.5 0.6-0.1 2.9-0.4 1-2.2-1.9-2.3-3.5-0.2-1.8 1.4-1.7-0.3-1.2-1.6-2.5-2.1-1.5-1.2-0.5-0.7-1.6-2-1.3 0-1 2-1.7 1-2.6 1.9 0.5 1.4-0.2 2-2.4 1.6 0.1 3.9 2.1 4.2 1.2 2.1 1 1.2 1.1 1.6 0.3-0.5-1.4 0.8-0.4 2.1 0 0.2-1.6-2-0.6-0.8 0.2-2.4-1.4-2.3-0.8-2.2-1.5-1.2 1.3-1.2 0-3.4-0.9-2.1 0.7-4.2 0.7-1.1-1.3-2.6 1-1-0.7-1.9-4.2-1.7-1.9-1.2-0.5-0.2-1.9 1.5-0.2 2.4 0.9 0.9-0.6-0.6-1.6-2.5-0.2-1.8 0.5-2.8-2.2-1.5-1.7-2.2-1.1-1.7-3.4-2.6-2.6 2.7 0.9 0.7-0.7 0.3-1.5 0.9-0.7-1-2.2 1.7-1.3 1.9-0.3 0.6-0.5 0.3-2 1.2-3.1 1.7-0.8 0.8-2.6-1.1-1.8 0-1.1 4.6-0.3 0.9-0.6 2.3 0.6 2.4-1.3 1.7-2.5 0.8-0.3 3-0.4 3.3 0.6 2.9-1 0.2-2.1 2-0.3 3 0.2 1-1 2.4-0.2 2.6 0.2 3-1.1 1.9-0.2 0.2-0.9 4.2 0.1 1.6-0.5 1.2 1.7 2.4 1.4 0.9-0.6 3.7 1.2 1.7 1.1 3.4-0.9 3.2-0.4 1.8 0.3 2.5-1.2 0.2-0.8-0.9-2.7-0.7-1 1.6-1.1 1.6 0.5 1.2 0.3 2.3 1.5 0.9 3.1-3.4 2.5 0.5 4.4-2.5 3.2-0.8 0.4z"
+                id="GR" name="Greece">
+            </path>
+
+            <path d="M611.2 521.2l2.3 0.4-1.1 0.9-1.2-1.3z m12.9 1.5l1.6-0.1 0.6 0.9 6.2 3.8 1.7 0.7 1.2 1.9-4.8-3.1-4.4-2.5-3.1-0.7-4.3-2-0.5-0.9 5.8 2z m-6.3-3.5l-0.8 0.1-5.6-0.1-1.7-0.3-0.5-1.3 1.8 0.2 0.5 0.6 6.3 0.8z m-5.3-2.3l-2 0.1-2.6-0.8 0.3-1.1 4.8 0.6-0.5 1.2z m-21.2-16.5l-0.3 0.4-3.3-2.7-0.3-1.9 2.8 2.7 1.1 1.5z m-4.9-6.5l-1.9-0.2 0.8-1.1 1.1 1.3z m-4.2 1.5l-0.9-0.1-1.7-4.2 0.4-1.2-0.8-2 0.9-0.3 0.5 2.1 1 0.9-0.1 1.6 0.7 3.2z m3.7-4.9l-1.5 0.3-1-1-1.3-0.2-1-1 1-0.7 0.5-1.2 3.3 3.8z m49.5-14.7l-0.7 1.5 1.6 2.7 0.2 1.6 1.3 0.4-0.4 2 1.2 1 3 1 0.9 1.5-2.3 0.1-1.7 0.4 0.7 2.6-1 1.6-2.2-0.2-1.3-2.2-1.1-0.8-2.2-0.2-2.6-0.7-1.8 0.2-1.8-0.3-2.2 1-1.6-1.3-1.7 0.6-2.8-0.7-2.5-0.1-2.7-1.6-1.5 1.2-3.2-0.3-2.7 3.2-1.8-1-1.7-1.8-1.8-0.4-1.2 0.6-0.3 5 0.1 1.4 1.9 1.3 3.4 3.2 1.3 4 2.3 2.4 2.8 2.3 1.7 2 5 4.1 2.4 1.1 0.4 2.2 1.5 1.6 3 2.5-0.3 1.6-3.5-2.8-2.6-1.5-3.1-2.8-3.8-1.1-2.7-1.2-3.3 0.6-1.8-0.2-0.7-2.2-5.8-3.9-4.3-4-0.8-1.3 3.1-1.2-1.4-0.8-3.9-3.4-1.2-1.6-0.1-4.1-0.8-1.7-2.9-2.2-1.1-1.1-2.1-0.7-1.4 0.9-0.3 1.9-3 5-1.3 0-3.1-4.2-0.5-1.9-1.3-3.9 0.7-0.5 0.5 0.6 3.3 0.8 1.4-1.3 1.1 0.5 3.6-0.1 0.6-0.3 1.2-2.2 2.6 2.1 1.5 0.7 1.3-0.5 2 0.7 1.6 0.2 1.2-0.4-0.8-2.3 0.7-0.7-1-1.1 2.1-1 2.1-0.6 0.4-1.1-0.2-2.2-0.9-0.9 0.4-1.4 4.4-1.7 0.7-1 2 0 0-1.7 0.9-0.8 2.5 0.6 2.9 1.2 1.7 1.2 0.9 1.3 2.7 1.8 2.2 2.2 1.3 0.7 2.5 0.6 1.3 1.3 3.3 0.9 4.9 0.2 2.5-0.9 1.1-1.4 2.9-0.4z"
+                id="HR" name="Croatia">
+            </path>
+
+            <path d="M671 437.8l1.5 0 1.4 2.3 2.2 0.8 0.9 1.1 2.3 0.4 0.9 0.7 0.7 1.7-2.2 2.3-4.6 1.1-1.9 2.4-1.2 1 0.1 1.7-1.4 1.4-0.7 2.5-1.4 1.5 0 1.2-1.6 2.7 0.2 1.3-2.1 1.5-1.5 4.7-1.3 0.9-2.6-0.3-2.5 2-1.4-0.5-3.2 0.9-1-0.5-2.8-0.3-1.1 0.2-2.9-0.4-1 0.3-1.5 1.6-2.2 1-1.6-0.5-2.1 1.3-2.9 0.4-1.1 1.4-2.5 0.9-4.9-0.2-3.3-0.9-1.3-1.3-2.5-0.6-1.3-0.7-2.2-2.2-2.7-1.8-0.9-1.3-1.7-1.2-2.9-1.2-1.9-2.2-1.4-3.3-2.4-0.1 1.8-1.7 2.5-0.5 0.3-1.8-1.1-1.6 0.5-0.7-0.4-1.5 2.2-1.2 0.5-1.4-0.5-1.1-2.6-0.7 1.3-1.4 2.5 0.9 3.9-0.3-0.7-2 0.6-1.9 0.6-0.7 2.1 0.3 4.2 2.7 1.6 0.7 4.7 0.1 7-0.4 0.6-1-0.5-1.3 0.4-1 1.4-0.7 4.3-0.4 2.4-0.6 1-1.5 0.7-0.2 3.5 1.4 5-2.5 1.6-3.5 1.8-0.4 2.7 0.1 2.5 0.6 4.6-0.7 1.4 0.9 2.2 2.2 0.6 0.2 4.3-1.1z"
+                id="HU" name="Hungary">
+            </path>
+
+            <path d="M329.3 352l0.7 1.1-2.2 0.4 1.5 5.2 0.6 0.9 0 4.1 0.9 3.2 0.1 2.5-1.3 1.8-0.9 2.1-0.2 1.8-1.6 2.1-1.5 0.9 1.6 1.5-2.8 0.8-2.6-0.3-1.3 0.8-3.9 0.2-3.7 0.7-1.7 1.8-2.5 0.7-1 1-2.3 0.8-1.7-1.2-1 0.2 0.2 2.1-2.9 1.3-1.8 0.3-1 0.7-5.8 1.4-1.9-0.5-3.3 0.8-1.2-0.1 1.6-1.8 1.6-1.3-3.8 0.6-1.4 0.5-0.6-0.6 3.2-1.9-3.3 0.1-3-0.4-0.4-1 2.9-2.3 2.4-0.9-0.1-0.9-5.1 0.1 0.5-1.1 1.7-1 1.8-0.1 1.5 0.6 1.1-0.8-0.6-1.7 3.9-2.3 3-0.3 3.3-0.7 3.3-0.8-2.4-1.2-2.3 1.8-2.6 0.3-1.9-0.5-2.1 0 2.7-2 1.6-1.7-0.8-0.7 2-2.3 3.6-1-1.6-1.3-3.9 0.2-1.4-1.3-2.6 0.1 0.4-1.2-3.5-0.2 0.5-1-0.6-1.1 2.8-0.6-0.2-2 1.9-0.8 2-0.4 0.1-1.1-2-0.2-2 0.4 0.8-2.1-0.8-1-0.4-1.7-0.9-1 2.8-0.2 1.2-0.5 4.7 0.1 1.9 1.3 1.7-1.2 4.8 0.9 0.6-2.6 3.8-1.5 1.2-2-3.7 0.4-3.6-1 0.7-0.8 2.1-0.7 1.9-1.6-0.3-1.1 1.7-2.7 3.1-0.8 2.4 0 1.6-0.9 1.2 2-1 1 0.6 0.9 1.2-0.8-0.5-2.1 0.9-1.3 1.9-0.3 0.6-0.8 3.2 1.7-2.5 1.5-0.5 0.7-1.9 0.9-1 2.3-1.2 1.6-2.2 0.8-1.6-0.1-0.2 0.8 1 1.4-2.9 0.9-1 1.4 2.7 2 0.2 1 2.1 0.4 0.8 0.7 2.9 0.3 2.4-1.8-0.5-0.9 1.9-1.6 1.3 0.5 1.8 2.7 1.4 0.5-0.1 1.5 2.9 0 2.1-0.4z"
+                id="IE" name="Ireland">
+            </path>
+
+            <path d="M240.3 177.9l1.2 0.1 2.8-0.8 2.2-1.1 0.5 1.1-2.5 2.4 2.2 0.9 2 0.3-0.1 1.7-1.7 1.6 0.8 0.3 3.5-0.5 1 2.3 1.5-0.2 2.4 0.3 1.4 1.2 1.2-0.3 0.5 0.9-0.7 1.2-1.5 0.6 2.1 2.9 0 1.1-1 0.7-1.3 0-1 2.1-1.2 1.1-2 1-2.6-0.5-1.1 1.5 0.2 1.4-1.3 1.3-5.3 2.1-2.5 0-2.6 0.5-6.2 2-4.7 2.8-2 0.7-3.1 0.3-1.8 0.5-5.9 0.8-2 0.5-1.2 1.1-0.5 1.4-4 1.1-3.8 0.7-6.4-0.5-5.4-1.3-4.5-0.3-3.8-2.5-2.9-1-3.6-1.8-2.3 1.1-12.4 0.3-0.8-0.5-0.2-2.1 1.1-0.8 1.3 1.2 5-1.1 2.7-1.9 2.5-2.7-3.5 1.2-0.2-1.9-0.9-1.1-1.7-0.8-0.3-0.8 1.1-0.9-2.3-1.2-2.6 0.1-6.6-0.5-3.8 1-2 0.1-0.9-1 0-1.1 1.5-0.3 1.8 0.1 4.1-0.6 1.4-0.6 0.8 0.5 3.6-1.1 1 0.2 4-0.2 4.4-0.1 1.1-0.8 0.4-1.1-2.8 0.9-3.7-0.5-0.6-0.9 3.6-1.6 2.7-0.8-0.8-1.3-3.2 0.2-0.7-0.8-2.6-0.4-1.8 0.3-0.9-0.5-2.4 0.6-5.2 0.9-3.3 0.9-3.1-1.1-2.4-0.2 1.4-1.4 2-0.1 1.7 0.8 0.1-3.2 1.2 0.3 2.8 1.1-0.7-2.8 2.2 0.1-2.2-1.6 0.5-1 1.6-0.3 0.8-1 2.4 0 2.3 1.1 0.2 1.4 2.1 0 1.1-0.8 0.9 0.6 0.5 1.2 1.2 0 0.4-2.2-3.6-1.1-0.3-1.4 3.7-0.1-1.7-1.1-3 0.2-1.4-0.4 2.6-1.2 3.4 0 1.4 0.2 2.3 1.7 2 0.6 3.1 2.3 2.2 0.9 0 1.2 0.7 0.8-2.3 3.4 0.9 0.7-0.8 1.6 1.9 0.6 1 1.7 0.9-1.8 1.5-1.9 2.3-1.3 0.9 0.1 0.9 1 0.9 0.1 1.4-2 0.7-4.3 1.7-0.9 1.8 0.7 1.3 1.6 2.6 2.1 1.8-0.5 0.1-2.3 1-1.3 2-0.3 2.2-0.8 2.3-0.2 1.7 1.4 1.2 1.4 1.7 1.1 0.7 1.8 0.9-0.3-0.2-1.7-1.4-2.6 0.3-0.9 4.1 0.2 3 2.3 1.3-0.3 3.6-2.8 2.3 0.8 1.8 0.1 2.8-0.8 0.8-1.2-0.9-2.3 0.6-0.4 2.6-0.6 2.7 0.1 2.6 2.1-0.1 1 1.6 0.8z"
+                id="IS" name="Iceland">
+            </path>
+
+            <path d="M602.6 595.6l-1.2 2.5-2.8 4.3-1.5 5.1 0.3 2 1.6 1.4-0.7 0.5 1.6 1.8 0.1 1.3-1.8 1.9-0.4 1.7 0.1 1.5-2.9-0.6-1.4 0.3-3.7-1.4-1.9-2.8-3-2-3 0-1.4-0.5-2.9-1.9-3-1.4-2.6-2.1-1.7-0.4-1.5-1-2.9 0-2.3-1.6-1.3-2.3 1.3-3.7 1.4-0.8 0.9-1.2 2.3 2.3 1.8-0.8 1.4-1.6 2.4 0.1 0.5 0.9 1.4 0.3 2.5 1.6 1.4 0.4 3.4-1 3 0.4 2.8-0.5 1.7-0.6 1.2-1 1.4-0.3 3.5 0.3 2-1.2 0.8 0.2 2-1.2 1.1 1.1z m-78.3-41.5l0.7 0.9 1.8 5-0.6 1.6-1.3 2 0.8 2.8-0.6 11.5-0.6 2.9-0.9 0.4-2.9-1.2-1.5 0.3-1.2-0.6-0.3 3-1.8 2.1-3-0.2-0.7-0.6-2.4-3.8-0.4-4.3 0.6-1.3 0-2.5 1.1-0.1 0-1.7-1.8-1.2 0-1.9 0.7-1.4-0.1-2.6-0.8-0.9-0.8-2.3-2.2-2.4 0.1-3.4 3.4 0.6 1.3-0.3 3.1-1.5 2.1-2.5 1.4-0.5 1.5-1.8 0.7 0.9 2.2 0.8 1.2 1.5 0.9 0.5-0.7 1.4 1 0.8z m8.3-30.8l0.3 1.4-2.5 0.5-1.1-1.1 2.9-0.2 0.4-0.6z m37.9-56.6l-0.2 0.9-3.3 2.3 0.7 1.4 2.3 0.7-1.6 2.1 0.3 1 1.3 0.2-0.4 2 1.9 1.1 1.5 1.3 0.2 1.3-1.5 0.1 0.7-0.6-2-2.2-2 0.9-3.2-0.9-2.1 2-1.5 0.4-1.7 1.1-3.2 1.2-1.8-0.4-1 0.7-0.4 3.1 0.8 0.6 1.5 2.6 1.7 1.1-0.7 1.9-0.9 0.7-1.3-0.5-0.3 1.7 0.9 4.5 1.4 3.2 1.2 1.3 2.7 2.2 2.7 1.1 5.1 3.7 2.8 1.1 0.7 0.7 1.8 2.8 1.6 3.2 1.8 5.1 1.3 2.5 2.3 2.9 4.8 4.1 4.3 2.9 4 1.9 3 0.3 7-0.4 2.6 0.7 0.4 1.2-0.4 0.9-2.9 2.1 0 1.7 1.5 1.2 7.1 3.2 7.2 2.6 2.2 1.4 2.7 2.1 6.4 2.9 1.1 1.4 4 3 1.8 2.4 0.5 1.8-1.5 4.4-1.6-0.5-1.9-1.3-3.1-5.4-5-0.5-2.9-1.3-0.7-1.4-2.3-0.4-1.3 0.9-1.4 2-1.5 3-1.5 4.3 0 1.7 1.1 1.7 2.9 1 2.4 1.5 1.6 1.6 0.3 3.8 0.9 2.1-0.9 1.3-1.9-0.3-2.5 0.7-1.7 1.4-0.7 1.4 0.5 3.4-0.3 1.4-3.3 2.5-1.6 2.5-1 2.3-4.3 0-1.1-1.5-0.1-2.2 0.6-1.3 1.6-0.7 0.9-2.8-0.5-2 1.1-1.6 2.8-0.7 0-2.9-1.4-1.3-1.4-5.1-2.4-4.3-1.4-3.8-1.1-1.8-1.4-1-3.7-0.3-4.6-2.6-0.3-1.1 0.6-1.1-1.1-2.8-2-1.7-2.6 0.6-1.2-0.1-0.1-1.5-2-1.3-2.7-0.2-0.7-0.7-2.6-4-1.7-1.7-2.3 0.1-3.9-0.9-2 0.7-3.2-2.6-2.8-0.9-5.7-5.3-1.8-2-3.5-2.2-2.3-3.2-1.8-1.2-4-1.4-0.3-1.3-3.1-3.1-1.7-1-1.3-2.1-2.5-0.5 0.1-2.7-1.2-3.5-1.7-2.2-1.2-5.3-0.8-1.5-1.8-1.1-4-1.2-5.7-3.4-1.1-0.1-3.4-1.3-2.1-0.2-2.6 1.2-3.1 3.2-2.5 3.4-0.9 0.6-3.4 1.2-3 0.5-0.2-1.5 2.1-2.6-0.3-2-3.3 0.6-5.1-2.4-0.9-0.9-0.1-1.4-0.7-1.3 1.4-2.5 0.8-0.7-0.5-1.7-3.2-1.4-1.5-3.1 0.8-0.4 1.9 0.2 1.7-1.3 1.2-0.3 0.8-2.5-1.8-1.6-1.7-2.6-0.9-0.6-0.1-1.6 2.6-1.7 1.4 0.7 2.4-0.5 2.6-1 3.1 0.8 2.4-1.4 1.6-2.2-0.6-1.5 3.4-2.9 1 0.6 0.2 2.5 2.4 2 2.2 0.6-0.4 1.2 1.3 1.2 0.9 1.4 1.2-0.7-0.7-1.7 0.2-1 2.2-2.6 0.5-1.1 0-2.9 2-0.1 0.8 2.1 2 0.8 2.9-1.1 0.7 0.1 1.4 1.9 1.2-0.3-1.5-3.5 0.6-1.8 1.3-0.3 1 0.9 1.9 0.2-0.5-1.8 0.5-3.1 3 0.3 0.9 0.8 2.1 0.4 1.1-0.4 1.2-2.2 1.4-0.6 3.4-0.3 3.1 0.2 4.7-1.5 0 2.3 3.1 3.5 1.1 0.5 8.6 1.5 4 0.3 2.6 0.5z"
+                id="IT" name="Italy">
+            </path>
+
+            <path d="M519.4 458.5l-1.2-0.1-0.1-1.7 0.5-1.5 1.1 2.5-0.3 0.8z"
+                id="LI" name="Liechtenstein">
+            </path>
+
+            <path d="M707.9 328.5l0 1.8-0.7 3.1 3.9 1-1.3 1-0.4 1.1-3.6-0.1-1 2-0.8 0.6-2.6 0.6-1.1 3 0.4 2.3-1 1.5-0.5 2 1.9 0.5 0.9 1.1 0 1-2 0.2-1.7-2.3-1 0.7-2.1 0.5-1.4 1.2-2-0.2-0.8 2.6-1.8-0.1-3.3 1.3-1.6-0.9-3.7 0.3-1.6 0.3-2.9-0.4-0.8-3-1.1-0.9-4.5-2.2-2.3-0.7-0.6 0.6-1.3-2 0-2.1 1-3.1-1.3-1.3-1.4-0.6-0.9-1.3-5.7-0.1-2.4-0.6-3.9-1.4-2-1.2-1.8 0.2-0.6-1.2 0.2-1.6-1.1-2.4-1.6-2.9-0.7-3.8 2.8-1.7 3.6-1.9 4.7-1.4 5.9 0.5 4.3-0.1 1 0.9 1.6-0.7 4.9 0.5 2.3 0.1 3.7 0.9 4.6-0.4 1.7-1.3 2.3-0.2 2.4 2.9 6.1 1 0.9 0.4 5.4 3.1 2.6 2.1 1.5 0.7 2.3 0.5z"
+                id="LT" name="Lithuania">
+            </path>
+
+            <path d="M475.4 411.7l0.3 2.2 1.5 1.6 2.8 1.1 0 1.4-0.9 0.9-0.6 3-1.3-0.7-2 0.6-1.4 0-2.1-1.2 1.1-1.7-2-2.4 0.2-1.7 1.4-2.5 1.3-1.3 1.7 0.7z"
+                id="LU" name="Luxembourg">
+            </path>
+
+            <path d="M711.6 301.2l1.9 0.3 0.5 1.1 3.2 1.7 0.7 2.2-0.8 1.6-0.3 3.1 1.8-0.3 1 0.6 0.3 1.2 1.7 2.1 1.3 0.8 1.9 4.2-0.3 1.7-1.3 0.2-1.4 0.8-2.5 2.4-0.4 1.7-3.1-0.1-3-0.4-1.1 0.3-1.2 1.5-2.6 0.6-2.3-0.5-1.5-0.7-2.6-2.1-5.4-3.1-0.9-0.4-6.1-1-2.4-2.9-2.3 0.2-1.7 1.3-4.6 0.4-3.7-0.9-2.3-0.1-4.9-0.5-1.6 0.7-1-0.9-4.3 0.1-5.9-0.5-4.7 1.4-3.6 1.9-2.8 1.7-0.8-2.8-0.6-5.5 0.1-2.8 1.9-1.6 0.8-1.2 0.6-4.5 2.5-3.6 2.4-0.4 3.1-1 3.5-0.9 1.4 1.9 4.9 3 1.2 1 2.3 3.4 4.4 1.8 3.1-0.6 3.6-2.4 1-1.1 0-1.1-1.2-4.7-1.1-2 0.1-1.3 8.4-2.8 1.5 1 1.1-0.7 5.6 2.5 2.3 0.5 0.6 0.7 3.5 2.7 2.8 1.1 4-1.2 3.4 1 1.9 0.2z"
+                id="LV" name="Latvia">
+            </path>
+
+            <path d="M751.4 483.2l-2-2.3 0.9-0.7-1.4-5-0.1-2.6 0.9-4.8-0.5-2.9-1.4-2.3-1.4-2.9-3-2.1-1.3-1.7-2.1-1.8-2.5-3.2-1.5-1.3-3-4.9-2.5-3-2.5-1.6-2.1 0 0.2-0.6 2.3-1.4 0.7 0.2 4 0.1 1.2-1 0.8 0.3 1.6-1 3.5 1 3.6 2.4 2.5 0.3 0.6 1.2 5.6 0.6 1.4 1.9 1.4 1 1.4-0.7 1.4 1.4 0.6 2.3-0.6 3.1 0.6 1.2 2.3 1.2 2.9 2 0 2.4 0.9 1.9 2 1.3 2.1 0.8 1 1.6 0.2 2.9 2.9 1.7-0.6 0.7-2.9 0.4-1.2-1.3-1.6 0.6-2.1-0.8-1.9-0.2-2.1-1-2.5 1.6 0.5 2.1 1 1.7-0.4 2-2.3 1.7 0 1.3-2.5 2.9 0.5 1.4-0.4 1-2 0.2-1.1 0.7z"
+                id="MD" name="Moldova">
+            </path>
+
+            <path d="M683.8 531.8l3.3 3.2 3.4 1.8 1.7 3.3 0.9 1-0.5 1.7 0 4.2-2 0.3-0.2 2.1-2.9 1-3.3-0.6-3 0.4-0.8 0.3-1.7 2.5-2.4 1.3-2.3-0.6-0.9 0.6-4.6 0.3-1.3-1.1-1.6 0.2-1.3-2.4-1.2-1.1-1.2-2.2-0.2-1.9-0.7-2 0.8-0.8-0.3-2.1 0.5-2.6 1.7 0.3 0.5-0.8 0.2-2.6 3.4-1.5 1-0.1 1.6 1.2 1.5-1.9 2.1-0.4 2.2-0.3 2-0.9 1.9 0.3 2.3-0.8 1.4 0.7z"
+                id="MK" name="Macedonia">
+            </path>
+
+            <path d="M642.4 512.9l0.4 1.2 2.6 1.7 1.9 2 3 1.6 2.1 0.5 5.3 2.9 0.1 1-1.6 0.4-0.2 0.7-2 0.4 0.9 1.5-0.2 1.3-3.4 1.2-0.8-2.1-1.1-0.3-3.2 5.3-0.8 0.3-2.5-0.2 1.6 1.9 1.7 0.8 0.3 0.6 0 3.1-2.1-1.2-0.9-1.8-3.1-2.9-3.5-2 0.1-1-1.6 0.1-1.2-1.9 1.1-1 0.1-0.8-1.1-1.6-0.6-3 0.5-0.7 1.7-0.2-0.1-1.5 0.5-1.7 2-1.7 2.3-0.2-1.4-2.2 3.2-0.5z"
+                id="ME" name="Montenegro">
+            </path>
+
+            <path d="M458 260.9l0.1 1.7-1-0.1-0.7-1.1-0.2-2 0.7-0.5 1.1 2z m-1.8-11.1l-1.8 0 0.3-1.3 0.9-0.3 0.6 1.6z m32.5-31.9l-2.3-0.2-1-0.8 2.8-0.8 0.7 0.5-0.2 1.3z m3.6-4.6l-0.3-0.9 2.7-0.6 0.9 1-3.3 0.5z m28-16.6l-4.2 0.3-0.3-1.1 2.1-0.8 1.2 0 1.2 1.6z m6.8-10.5l-2 0.3 0.1-1.1 1.8-0.3 0.1 1.1z m4.1-5.7l-0.1-1.1 1-1.2 1 0.4-0.3 1.5-1.6 0.4z m3.1-24.9l-1.5 0.7 1.1-2.6 1.1-0.7 0.4 1.5-1.1 1.1z m8.5-5.2l2.3 0.2-0.5 0.8-1.4 0.3-1.1 0.8-1.9 0.2-1.9 1-1.6 0.3-0.4-1.2 0.9-1 1.3 0 1-1.2 3.3-0.2z m12.4-9l2.2 2.1-0.3 1.5-1.2 0.7-3.3 0.1-1.6-0.9-2.2 0.6-1.6-1.1 1.6-1.4 2.6 0.4 0.5-1.7 1.7 0.3 0.4-1.5 1.2 0.9z m6.3 5l1.3-1.1 1.5-0.4 0.6-2.2 1.8-0.5 1.6 1 0.9 1.2-0.2 1.1-3.1 1.2-2.5 2-1.6-0.2-0.7 0.7-2.4 0.6-1.7-0.9-0.9 0.9-1.4 0.4-1.2-0.4-2.9 1.5-2.8 0.3-1-1.1 3.1-1.9 5.1-0.6 2.9-2.3 0.4-2.5 0.6-0.9-1.3-0.6 0.2-1.7 2.4-1.7 1.3-1.4 1.5-0.3 0.9 1.1-1.1 1.4-1.6 1.2 1.1 1.5 0.6 2.4-1.4 2.2z m15.5-13.7l1.4 0.8 1.5-0.3 2.3 0.7 1 1.5-1.2 0.9 0.4 1.7-3.8 0.5-0.9-0.5-1.5 0.9-1.4 1.4-2.2-0.6 0.2-1.6-0.3-2.3 0.7-0.5 2.9 0.3-0.4-2 1.3-0.9z m124-2.6l-1.8 0.4 0.2-1.8 1.6 1.4z m-92.4-3.8l-1.3 0.4-1.8-0.3-0.7-1 2.2-1 1.7 0.3-0.1 1.6z m-15.1 0.3l1 0.7 1.5-0.5 1.1 1-1.4 0.8-0.8 1.2-2 0.1-2.9 2.3-0.1 0.8-5 0.5-1.5-0.3 1.7-2.7 1.7 0 1.7-0.8-0.3-0.8 2.5-2 0.6-1.7 1.4-1 0.8 2.4z m4.8-2l1.4 0.2 1.3 1.6-2.3 0-2-1.8 1.6 0z m37.3-4.3l0.5 1.1-3.3 2.1-1.6 0.1-2.1-1.2 0.1-0.8 1.9-0.9 0.9 0.1 2.7-1.4 0.9 0.9z m3.9-0.3l-1.8 0.5-1.7-0.9-0.2-1.6 0.8-0.3 3.2 1.2-0.3 1.1z m-6.4-3.2l-0.4 0.8-2.7 2.1-2.2 0.7-1.8-0.3-0.8 0.6-2 0-2-0.6-2-1.2 3.5-0.1 0.6-0.5 1.5 0 2.8-0.4 1.1 0.3 2.1-1.4 2.3 0z m60.6 23.5l-1.7-1.3 0-0.7 3.2-2.9 0.7-1-2.6-2.7-3.6-0.8-4.2-1.1-4.2-2-1.8-1.2-2.9 0.3-3.5 1.6-0.9 0.2-3.8-0.4-2.1 0.2-1.7 1.8-2.6 1.7-0.7 3.8-0.6 1.7 0.7 3.2-1.5 1.4-2.1 0.3-1.6 1.3-0.5 1.7-1.3 0.6-2.6-0.8-3.9-0.8-3.6-1.1-1.5-0.1-1.2 1.2-3.7 0.9-1.8 0.1-0.8-0.5-6.9-0.6-1.4-1.8-7.9-5.5-3.6 0-1.9 1.1 1 1.4-0.5 0.6-1.9-0.4-2.1 0-0.5 0.4-5.1 0.2 2.6 1.6 0.1 2-1.4 1.9-1.6 0.9 2.9 0.9-2.5 1.6-1-0.1-8.6-1.7-2.9-0.1-4.1-0.9-2.1 0.5 0 0.8 0.8 3.6-2.1 3.1-6.3-1.8-1.5 0.9-3.7 1.9-1.5 3.5-1.1 1.1-2.7 0.6-0.5 1.1 2.7 2.3 1 1.3-0.2 1.4-1.5 1-3.3 2.9-3 2.9-1.3 0.9 1 2.5-3.3 1.5-2.4 0.5-3.8 0.4 1.4 3.9-0.6 2.7-0.1 4.7-3.3 5-4.4 5 4.7 1.6 1.1 2.9 0.1 1.2-1.7 2.2-7.1-0.9-3.3 0.4-3.3 1.5-3.5 3.8-1.2 1 0.5 1.5-2 2.8 2.7 4.1-1.1 1.1 0.6 2.8-0.1 1.8 2.5 4.4 0 1.6-0.9 6.4 3.8 2.1 1.2 0.4 3.4 2.7-0.4 1.9-1.3 2.5-2.4 0-1.8 0.6 0.3 1.6 3.1 4.9 0.5 1.4-0.6 2.1-0.1 2.8-1.9 2-1.3 0.8-2.6 0.7-0.9 2.4-1.5 1.5 1.6 4.4-0.6 3.9-0.7 1.4-1.9 0.2-1-1.8-0.4-1-6-1.2-2.3-2.9-0.4-3.1-2.5 1.2 0.6 1.1 0 2.3-2.5 3.9-1.1-0.2-1.3 0.8-1.3 0.2-0.6-1-1.1 0.2-0.8 1-3.4 1.3 1 0.7-1.7 1-0.6 0.9-2.7 1.5-4.3 3.9-2.3 1.1-1.6 1.2-1.4 0-1.8 0.9-4.6 0.9-3.1-0.4-2.1 0.4-1.1-0.7 0-1.1-1.2-0.1-0.8 1.2-1.6-0.4 0.7-2.4-3.1-0.1-3.8-1.6-0.9-0.8-3.1-1.3-1.4-1.4-0.8-1.6 0.3-3.6 0.6-0.5 2.7 0.8 2.8 1.3-0.1-1.2-2.4-1.6 0.8-1.8-0.2-1.6 2.6-2.1-0.4-0.7-3.4 2-3.1 0.9-1.1 1.2-1.1 0.6-2.1 0-0.5-0.9 0.5-3.3 1.2-2.7 1.8-1 3.4 0.7 1-1 1.3-0.2 2.4-1.2-4.1 0.5-1.2-1 2.8-2.4 1.2-1.4 0.3-1.5 2.2-1.7 1.8-0.8 1.6 0.6 1.5-1.9-5 1.1-2 1.1-2.7 2.8-0.2 1.1-2 0.9-1.5 1.5-0.6 1.2-2.8 1.9-1.4 0.3 0-2.4 1.1-2-0.3-1.6 3.2 0.2 1.8-1.2-2.7 0-1.4-0.7-1.2-1.5-0.6-1.9 1.1-1.8-2.2-2-0.5-2.2-0.1-2.5 0.9-0.5 2.1 0.4 2.4-0.2 5.2-0.9 3.4 0.6 1.4-0.1 2-0.7 1.9-0.1 2.2 1.2 0.7-0.5 5.4-1.2-3-1.5 0.1 1.4-3.6 0.3-3.2-0.6-1.3-0.6-0.5 1.3-1.1 0.3-3.4-0.5-4.7 0.3-2.1 0.7-1.4-0.1-2.5-1.2-0.9-0.9-0.3-2.6 2.8-0.3-1.9-2.3-1.9-1.3-0.2-1.4 0.7-1.3 1.5 0.2 3.8-0.1 6 1.5 4.9-0.3-0.8-0.6-2.9 0-7.2-1.3-2.2 0.1-1.2-0.3-0.7-1 0.4-1.9 1.1-0.4 1.2 0.6 1.9-2.3 1.9-1 2-0.5 2 1 1.3-0.1 4-0.8-1.3-0.5-3.5 0.5-0.2-0.5 2.4-2.3 4.2-0.4 5.9 0.5 2.3 0.8 0.4-1.8 5.3-1.4-6.8 0.8-1.8-0.5-4.9 0.7-0.1-1.3 2.3-2.3 4.1-0.9 1.9-1.1 3.1-0.2 2.6 0.3 1.2 1.3 4.4 1.7-5.1-4 1.1-1.7 3.3-0.4 0-1.6-2.2-0.3 0-1.2 1.9-0.9 2.8-0.7 3.2 0.8-0.6 1.3 0.9 0.5 1.7-1.5 3.9-0.8 2.8 2.1 0.5 1.1 1.1-0.5 2.7-0.5 2.7 0.3 1.8-0.2-1-1.3 0.4-1 2.4-1 1.7-0.3 2.6-1.2-0.8-1-1.4-0.3 2.9-1.5-1.7-0.6-2.4 0.9-1.6 0.9 1.5 1-1.1 1.1-6.1 2.8-3 0.8-1.4-0.1-1.9-2.5-2.1-0.1 1.3-2.2 1.6-0.8 1.4-2.3 2.3-1.3 3.2-3.3 2.8-1.1 0.9-1.1 3-1.5 3-0.9 0.3-1-2.7 0.9 0.4-2.1 1.4-1 6.8-2.8 1.6 1.3 2.1-0.2 2.3-1.6-0.3-0.7-2.1 1-1.5 0-0.4-0.6-2.1-0.5-0.3-1.2 1.3-2.8 3.4-3.3 0.5-1.3 1.3-0.7 1.7 0.2-0.2-1.1-2.1-1.1 6.2-1.1 3 0 2.4-0.9 0.5-1-6.4 1.1-2.4 0.1-0.8-2.7 0.2-1.5 1.9-2.2 1.4-0.2 1.7-1.1 0.1-1.5 2.3-0.8 2.1-2.1 2.3-0.5 1.4 0.2 3-0.7 6.6-0.1 0.1-0.6-6.3-0.3-3.5 0.2-0.4-0.6 2.8-2.7 2-1 1.7 0.2 1.8 1 0.6-0.8-1.6-1.5 0.7-1.5-2.6 1.1-1.8 0.3-0.9-2 2.3-2 2.3 0 0.3-1.5 3.3-1.3 1.9-0.2 1.5-0.6 0.6 0.4 0.9 2.3 1.5 0.3 1-0.8-1.3-1.1-1.1-1.8 2-1.5 2.3-0.2 2.6 0.7 2.3-0.2 0.9-1.2-6.3-0.1-0.8-0.9 1.2-1.2 2.2-0.8 2.5-0.1 2.4-1.4 0.9-1.1 0.3-1.6 1.4-1.3 3.9-0.8-0.6-2.2 0.8-1.4 1-0.6 2.1 1.5 1.8 0.5-0.5-1.4 0.3-1.2 2.1 0 1-0.5 0.3-1.3 1.4-1.2 4.7-0.7 0.5 2.5 0.7 1.4 0.8-2.8 1.6-2.1 2.4-0.9 0.9 1.1-0.4 4.3-1.9 2.4 1.5-0.2 2.5-2.3 0.6-1.2-0.7-2.8 0.6-0.8 3.5-0.1 1.9 0.4 0.7-1.5 1.7-0.1 1.8 0.9 4.1 1.4-1.3-2.2-1-0.8-2.1-0.5-2.2-1 1.2-1 2.4 0.3 2-0.8 2.2-0.2 1.1 0.6 0.8-1 2.5-0.5 2.8 0.9 1.6 2.7 2.4 1.4-0.1-1.4 0.6-1.9 2.3-2 2.1-1 1.4-0.1 2.9-3.4-1.6-1 3.3-2.3 1.1-0.1 5.4 2.1 1.5-0.9 3.4 0.8-0.8 1-3.1 2.9-0.5 0.8-0.1 2.2-1.2 1.4 0.8 1.4 1.6-0.4 1.8-1.3 0.1-1.3 4.3-3.7 1.9-2.1 2.3-1.7 1.5-0.3 0.9 1.1-0.2 1.5 0.2 4.3 3.7-1.4 1.1-3 1.1-0.7-0.8-1.4 0.4-1.1 2.3-1.3 2.2 0.4 3.3 0.3 2.7 0.8 0.2 1.4-0.8 0.9-2.9 1.1 0.7 0.7 2-0.6 0.9 5.5 0.9-2.1 1-0.7 0.1-1.5 0.8-1.9 1.8-1.3 2.7 0 3.3 1.7 2.5 0.3 1.8 0.8 2.4-0.7 2 1 0.1 1 3.8 0.3 3.8 1.6 0.6 1.6-4.4 1-1.8 1-3.3 0.4-11.5-0.7 0.4 0.7 8.3 1.5 0.6 0.5 0.4 2.1 1.8 0.7 3.5-0.2 0-1.3 3.2 0.6 3.8 0.2 1.1 1.7-0.3 1.5-2.4 0.1-4.2-1.4-0.4 2.3-1.1 0.9-5.8 1.3-1.1 2.6-2.2 1z m-40.5-27.7l3.7 0.5 2.2 1.4-3.6 0.5-1.7-0.1-2-1.8 1.4-0.5z m-339.6 3.9l-0.1-1 4.5-1.5 1.9-1.4 3.4-0.5-0.2 1.8-3 0.8-3.3 0.5-3.2 1.3z m269.2-45l-1 0.5-1.9-0.8-0.8-1.2 3-0.1 0.7 1.6z m9.6-50.1l3.9 0.2 2 1.9 1.4 2 1.3 0.2 4.7-0.4 3.4 0.7-1.4 1.1 0 1.2 5.5 1.3 1.8 0.1 3-0.4 6.5 1.7-6.6 1.1-2 1.9-3.7 1.3-2.6 0-2.3 1.3-2.2 0.1-1.5-1.9 2.2-1.3-3.1-0.4-1.2 0.5-10.5 1-1.1-0.2-0.9-1.2 2.8-0.6 0.2-1.1 2.5-2.4-5.5-1.6-2-1.2-0.8-1-2.5-0.9 0-1-1.7 0.1-0.4-1.2 10.8-0.9z m46.2-0.6l-1.3 0-2.6-0.8 0.7-1 3.2 1.8z m-137.6 0.4l3.5 1.5 1.9 0.5 2.3 1.5-0.7 0.9-2-1.2-3.6-1.3-2.3-0.2-3.3-2.5-1.5-0.8-0.8-1.7 3.6 0.7 1.6 1-0.3 0.8 1.6 0.8z m155.6-3.4l2.7 0.1-0.1 0.6-7.3-0.4-3.8 0.7 0.4-0.9 2.4-0.6 3.7-0.1 2 0.6z m-109.9-10.8l1.4-0.6 5.4 0.8 2.5 0.9 1.3 1-1.9 1.9 1.7 1 3.2-1.7 2.7 0.3 1.7 0.9 1 1.1-0.2 1.4 3.1 1.1 3.4-0.2 2.4 0.3 1.5 1 1.9-0.2 0.2-0.8 2.6 0.2 3 0.8-2.1 0.9 2.1 0.8 3.4 0.6 2.1 0.6-1.5 1.4-3.2 0-2.9 0.3-5.4 0.3-3 1.6-2.5 1.8-0.4 1.6 0.6 1.1-1.5 0.5-3.3 0.2 0.2 3.9-0.8 1.9-3.3 0.3-1.8 1.1-1.2 2.1-2.1 2 1 1-2.1 1.9 1.2 1.8-1 0.6-2.1 0.3-2.2-0.3-2.3-1.1-6.7-2.2-4.3-2.4-6-1.3-3.3-1.6-1.5-1.3 0.3-1.4 2.8-0.2 2.9 0.7 1.7-1.9 11.3-1.1 7.3-0.2-1.9-1.3-6.9 0.7-4.3-0.2-2.1 0.1-2.1 0.6-4.3 0.5-4.6-0.1-2.9-1-1.3-2 1.2-0.6 3.8 0.1 6.4-0.9 3-0.8 3.1-0.1-0.5-0.8 1.8-0.4 2.4-0.2 5.6 0.1 0.2-0.6-2.6-0.6-0.9-0.6 2.6-1.8-2.2 0.1-3 1.3-1.8 0.5-4.6 0.3-1.7-1.6 0.9-1.2-0.8-0.7-1.3 1.3-2.1 0.3-2-0.9-1.9 0.5 0.7 1-0.6 1 2 0.9-2.3 0.6-2.1 1.1-3.9 0.2-4.5 0.1-4.6-1.6-2.2-1.2-1.8-1.2-2.7-0.6-1.7-0.8-0.4-1.1-1.4-0.7-0.7-1.2 4.4 0-1.7-1.8 0.5-1.6-3.4 0.1 0.4 0.9-2.2 1-1.2-0.2-2.3-1.9-2.4-2.4 0.4-1.3-1.2-1 1.5-0.7 2.5 0.9 0.8-0.7 3.8-0.4 4.6 1.1 0.4-1.1 4.2 0.5 3-0.6 5-0.3 2 0.4-1 1.1-6.3 0.3-4 1.3 5.7-0.2 1 0.1 0.6 1.1 3 0.5 2.3 1 0.8-0.5-0.6-2.2 4.4-2.3 2.1 0.4 2.1 0.9 4 3 2.3 1.9 6.1 2.8-0.7-1.5-4.2-4.5-0.9-1.7 0.9-1.7 1.1-0.5-0.4-1.3 1.3-0.5 2.4 0.3 2.6 1.3z m134.7-2.3l-8 0.4-0.9-0.3 15.7-1.4 1.7 0.5-8.5 0.8z m-118.9-1.9l-1.8 0.6-3.2-0.5 0.9-0.7 4.1 0.6z m18.6 0.5l5.6 0.1 1.5 0.9 5.5 1.1 0.8-1.5-0.6-2.2 1.9-0.1 1.7-0.6 3.1 0.3 0.2 1.3-0.5 1.4 2.1 0.1 2.8-0.3 1.9-1.1 2.4 0.1 1.3-0.6 1.4 0.6 8.2 0.7 3.3 0.6 5.2 0 3.7 0.2 2.8 1 1.1 1.7-0.9 0.4-6.7 2-1.6 0.7-2.4 2.3-4.1 0.7-3.5-0.3-3 0.7-1.9 1-1.5 0.2-9.2-0.4-2.2-1.1 1.2-0.9-8.2 0.3-9.2-0.1-1-0.5-5.6-0.5-4.2-1.1 1.1-0.5 5.8 0-0.7-1.5-2.9-0.1-1.8 0.4-4.3 0.2-4-0.1-1.9-0.3-3.8-1-0.7-0.8 2.8-0.4 2.1-0.7-4.5-0.3-1.9-0.3-0.6-0.9 5.9-0.2 5 0.8-0.7-2.2 3.9 1.1 0.2-1.1-2-1.4 2-0.1 4.5 0.7 5.1 1.6z"
+                id="NO" name="Norway">
+            </path>
+
+            <path d="M666.8 348l0.6-0.6 2.3 0.7 4.5 2.2 1.1 0.9 0.8 3 2.1 5.1 2.9 4.9 1.7 3.6 0.9 3.9-0.6 1.6-4 1.7-0.8 0.5-2 2.7 0 1.2 3.3 1.3 1.8 1 0.8 1-0.2 2.4 0.4 1.1-0.8 1.5 0.1 1.3 1 1.6 1.2 1.8 0 1.3 2.7 2.8 1.9 2.8 1.4 1-1.2 1.3 1.8 2.6 0.1 1.3-1.2 1.8-3 0.5-4.8 4.6-2.4 2.7-3.4 4.4-0.6 1 1.5 3.8-0.1 1.8 1.9 1.4-0.4 0.9-3.3-0.7-0.9-0.2-5.7-1.9-0.8-1.4-1-0.7-3.1-1-3.6-0.3-3.2 0.2-1.4 1.5-1.8-0.1-2.4-1-3 0.1-2.3 1-1.1 2.1-1.7-0.6-1.4 0.4-0.6-2.7-1.4-0.4-3-3-2.2 1.3-1 1.7-2.2 0-0.5-1.5-1.3-0.2-0.5-1.6-2.8-2.2-0.5-1.8-2.7-0.8-0.5 0.2-3.5-1.8-1.8 1-1-0.6-2.7-2.3 1.4-0.5 0-1.6-3.6 0.7-3.3-1.9-2.2-0.6-1.3 0 1.7 2.4-3.5 2.5-1-0.4-1.5-1.9-2-1.8-1.6-0.9 1.6-1.2 0.7-1.1-1.8-1.2-3.2 0.7-0.8-0.9-2.7-1.1-4.6-1.1-1-1.1-0.3-1.1-3.3-0.8 0.1 1.9-2.1 0.4 1.6-3.6 0.4-2.3-1.6-3.2-2.2-0.9 0-1.6-1.9-3.1 0.8-1.1 0.7-2.6-0.8-1.9-2-2.3 0.6-2.6-1.4-1.7-3.3-2.1-1.6-1.4 0.7-1.6 1.9-1.8 0.3-2.7-1.8-4.1-0.7-2.6 2.8 0.9 0.7-0.5-0.1-1.8-4.1-0.7-0.1-1.2 2 0.4 3.8-1.4 6.5-1.8 6.9-1.7 3.3-0.5 2-2.2 1.9-1.7 3.8-0.7 1.3-0.8 2.8-1.2 6.6-1.2 5.5-0.4 5.4 2.3-0.9 0.3-3-1.2 2.2 3.5 1.1 1.2 2 0.9 1.7 0.3 4.9-0.5 2.2-1.1 0.5 0.1 6.6 0.4 22.9 0.9 7 0.1z"
+                id="PL" name="Poland">
+            </path>
+
+            <path d="M306 611.9l-1.2 0.2-4.5 2.5-1.3 0-2.5-1.1-4.5-0.4-1.5-0.3-1.8 0.7-1.5 0-1.1 1-0.8-0.3 1-2.1 1.6-4.1 0-2.6 0.4-2.2-0.3-2.1-0.7-1.4 1.1-3.5 0-1.8-0.9-2.3 2.8 0.3-1.7-1.4-1.5 0-2.3 0.9-1.5 0.1 0.1-1.4-0.5-1.8 2-0.6 1.6-1.7-0.3-1.6-1.2 1.1-0.6 1.4-2.8 0.7-1.5-0.5 0-1.9 0.6-1.7 0.4-2.3 0.8-2.2-0.2-1.4 1.6-1.3 1.4-1.9 4.3-8.9-0.6-1 1.7-6.6 1.2-2.3 0.6-4.3-0.2-2-0.7-2-0.7-4.3 0-1.4-0.9-1.8 0.1-1 1.4-1.7 1.2-1.1 1.9-0.9 4.1-1 1 1.8-1.3 1.9 0.9 1.3 3-1.1 2.9-0.1 2.2 0.8 1.4 0 2.5-0.9 0.8-1.4 1.5 0.5 5.2 0.1 0.7 1.1 0.1 3.1 1.9 0.2 1.9 0.9 0.3 1.1-2.5 2.5-2.1 1.1-1.6 1.4-1.1 1.6-2 1.6 0.8 2 0.4 3.5-0.4 3.8 0.4 1.1-2.7 2.1-0.2 1 1.4 1.3 0 1.9-1.7 3.5-1 0.5-4.3 0-1.1 0.3 1.1 1.9 1.4 1.2 0.3 1.9 1.6 3.2 1.7 0.4 0.5 0.8-0.6 2.4-3.2 3-0.9 4 3 4.3 1.7-0.2-0.7 2.4-2.1 0.6-3.5 4.3-0.8 2.2 1.1 6.4z"
+                id="PT" name="Portugal">
+            </path>
+
+            <path d="M751.4 483.2l1.5 1.6 1.8 0.9 4 0.9 0.6-1.2 0.9 0.3 1.6-0.5 2.2-1.3 2.2-0.2 2.2 0.7 1.9 1.8-0.2 4.3-0.7 2.1-6.2 1.3 0.4-0.6-0.7-2-1-1-0.9 1.1 0.6 1.7-0.7 1.4-0.2 2.4 0.6 0.9-1.6 2-0.5 1.3 0.8 4.8-0.4 3.7-2 0-2.6-0.5-2.2-0.7-2.5-2.6-1.7 0.5-0.9-0.6-3.2-0.4-4.5-2.3-2.9 0.4-4.4 0.9-3.2 1.2-3.3 2.1-1.2 1.6-1.6 0.9-2.3 0.6-4.2-0.2-9.4-1.7-2.5 0.5-3.5-0.4-5.3-1-3.9-0.3-3.7 0.6-0.9-1.8 1.8-2-4.3-2.5-0.8-1.2-1.6-0.9-0.5-1.7 2.5-1.8-3-1.6-1.7 0.5-1.7 1.8-1.3 0.2-1-1.2-1.5-0.7-3.5-0.7-1.5-1.2-2.1-0.6 0.2-0.6 1.8-0.8-2.3-1.1 1.5-2-0.9-1.1-2.6-0.8-3.8-1.9-2.2-1.8-0.4-4.1-0.8 0.2-1.8-2-2.9-1.7-1.7-2 3.2-0.9 1.4 0.5 2.5-2 2.6 0.3 1.3-0.9 1.5-4.7 2.1-1.5-0.2-1.3 1.6-2.7 0-1.2 1.4-1.5 0.7-2.5 1.4-1.4-0.1-1.7 1.2-1 1.9-2.4 4.6-1.1 2.2-2.3 2.1-0.9 0.9-1.2 0.7 0 2.7 1.5 2.7-0.1 6.9 1.4 3.6-0.7 1.2 0.3 4.3 3.3 2.1-0.5 1-1.2 3.4-1.3 8.4-1.2 1.1-1.9 0.2-1.4 3.7-0.8 2.1 0 2.5 1.6 2.5 3 3 4.9 1.5 1.3 2.5 3.2 2.1 1.8 1.3 1.7 3 2.1 1.4 2.9 1.4 2.3 0.5 2.9-0.9 4.8 0.1 2.6 1.4 5-0.9 0.7 2 2.3z"
+                id="RO" name="Romania">
+            </path>
+
+            <path d="M651.6 473.1l1.7 2 2.9 1.7 1.8 2 0.8-0.2 0.4 4.1 2.2 1.8 3.8 1.9 2.6 0.8 0.9 1.1-1.5 2 2.3 1.1-1.8 0.8-0.2 0.6 2.1 0.6 1.5 1.2 3.5 0.7 1.5 0.7 1 1.2 1.3-0.2 1.7-1.8 1.7-0.5 3 1.6-2.5 1.8 0.5 1.7 1.6 0.9 0.8 1.2-0.9 0.7-0.2 1.9-2.1 1-0.4 2.3 0.6 2.4 2.4 3.9 2.9 1.5 1.3 1.6 1.6 1-0.3 1.4-1.6 1.7-1.1 1.6-1.8 0.1-1.1 0.6 0 3.3 1.4 2.9-1.1 1.7-1 0.3-1.4-0.7-2.3 0.8-1.9-0.3-2 0.9-2.2 0.3-0.6-1.3 1-0.9 1.2-3.2 0.2-1.2-4.7-1.2 0-1.3-2.2-1.2-0.2-0.7-2.3-2.1-2.8-1.3-1.1-1.3-1.7 0.9 0.6 1.6-0.4 1-1.7 1.3 0 1.4-1.5 0.5-0.1-1-5.3-2.9-2.1-0.5-3-1.6-1.9-2-2.6-1.7-0.4-1.2 0.7-0.8 2.5 0.3 0.2-2.2-3.2-3.7 0.5-0.8 3.1 0.1 0.3-0.8-3.3-2.9-2.4-1.2-0.5-0.8 0.1-2.6 1.6-2.7 0.5-2.5-1.5-0.8-2.8 0.7 1-1.6-0.7-2.6 1.7-0.4 2.3-0.1-0.9-1.5-3-1-1.2-1 0.4-2-1.3-0.4-0.2-1.6-1.6-2.7 0.7-1.5 2.1-1.3 1.6 0.5 2.2-1 1.5-1.6 1-0.3 2.9 0.4 1.1-0.2 2.8 0.3 1 0.5z"
+                id="RS" name="Serbia">
+            </path>
+
+            <path d="M674.8 427.7l-1.5 3-0.8 2.9-1.6 1.7 0.1 2.5-4.3 1.1-0.6-0.2-2.2-2.2-1.4-0.9-4.6 0.7-2.5-0.6-2.7-0.1-1.8 0.4-1.6 3.5-5 2.5-3.5-1.4-0.7 0.2-1 1.5-2.4 0.6-4.3 0.4-1.4 0.7-0.4 1 0.5 1.3-0.6 1-7 0.4-4.7-0.1-1.6-0.7-4.2-2.7-2.1-0.3-2.3-2.9-1.6-2.9-0.1-0.8 0.9-2.4 1.1-2.8 1.4-1.2 3.6 0.5 1.7-0.2 1.5-0.7 2.1-1.9 1.6-0.8 0.7-3 2.3-1.2 0.6-0.8 2-1.5 2.8-0.3 1.3 0.2 0.5 1.5 2.2 0 1-1.7 2.2-1.3 3 3 1.4 0.4 0.6 2.7 1.4-0.4 1.7 0.6 1.1-2.1 2.3-1 3-0.1 2.4 1 1.8 0.1 1.4-1.5 3.2-0.2 3.6 0.3 3.1 1 1 0.7 0.8 1.4 5.7 1.9 0.9 0.2z"
+                id="SK" name="Slovakia">
+            </path>
+
+            <path d="M605.2 467.1l-2.5-0.6-0.9 0.8 0 1.7-2 0-0.7 1-4.4 1.7-0.4 1.4 0.9 0.9 0.2 2.2-0.4 1.1-2.1 0.6-2.1 1 1 1.1-0.7 0.7 0.8 2.3-1.2 0.4-1.6-0.2-2-0.7-1.3 0.5-1.5-0.7-2.6-2.1-1.2 2.2-0.6 0.3-3.6 0.1-1.1-0.5-1.4 1.3-3.3-0.8-0.5-0.6 1.7-1.1 1.5-0.1-0.2-1.3-1.5-1.3-1.9-1.1 0.4-2-1.3-0.2-0.3-1 1.6-2.1-2.3-0.7-0.7-1.4 3.3-2.3 0.2-0.9 5 0.9 5 0.7 3-1.2 1-1.3 1.9-0.7 2.6-0.2 2.8 0.2 2.3-1.1 1.6-0.2 2.4 0.5 0.1-1.9 1.3-0.9 2.4 0.1 1.4 3.3 1.9 2.2z"
+                id="SI" name="Slovenia">
+            </path>
+
+            <path d="M591.1 319.4l-1.5-0.3-0.3-3.8 2-4.6 1-0.3 2.7-6.5 0.5 2-1.5 3.2-0.3 2.2-0.6 0.5-2 7.6z m26.2-22.7l-1.3 1.4-1.4 0.5 0.1 3.3 1.5 1.2-1.3 0.5-0.7 1.8-1.8 0.6-1.5 1.6-0.3 1.6-2.1 0.9 1.2-2.4-0.9-0.7-1.4-2.1 0.4-0.9-0.6-3.2 2.7-2.9 1.3-1.1 3.3-1 1.3 0 1.5 0.9z m36.8-113l-2.7 0.3-2.2-0.6-1 0.3-3.8 0.3-1.1 0.7-2.1-0.7-2-1.2-2 1.1-1.6-1-1 1.9 0.1 1.8-1.8-0.2 0.3 1.1-2 1.5-3.2 0.2 0.7 1.5-0.9 0.7 1 1.7-1.3 1.5-2.1 1.9-0.5 1 1.7 1.1 1.6 2.5 1.6 1.1-0.4 1.2-2.1 1.1-2.2 1.7-2 4.3-3.1 1.3-2.4 1.6-3 0.7-2.4 2.2-1.8-0.8-1.5 0.5-1 1.8-1.9 1.5-2.4-0.2 0.2 0.8-2.3 0.3-1 4-2.2 0.6-0.7 0.8-0.9-0.6-2.1 0.1 1.6 2.8-1.7 1.8-1.2 0-1.5 0.7-1.9-0.8-0.2 1.2 0.8 1.3 1.3 1-0.1 1.3-0.9 4.3 0.7 1.8-2.1 1.2-0.8 1.1 0.6 1.7-0.1 1.1 0.7 1.9 0.8 4.2 1.1 2-0.1 1.6 1.3 0.9 2.1-0.1 1.4 1.6 3.1-0.8 0.8 1.1 2.9 2.2 1.7 0.3 1.9 1.2 0 1.5 2.9 1 1.9 2 0.8 1.6-0.1 1-2.6 1.5-1.3 1.3-3.8 2.1-0.6-0.1-2.1 1 2 0.6 1.4-0.7 2.6-0.6 1.4 1.2-2.2 0.5-0.3 1.6-0.8 1-2 0.7-2.8 1.5-4 1.4-1.1 1.2-2.6 1-1.3 0.8-3.8 0-3.7-0.1 1 0.9 0.9-0.2 3.6 0.4 1.7 1.3-3 0.9 1.7 3.2-0.7 0.8 0.4 3.5-1.1 0.1-0.3 1.4 0.6 2.5 1 2.1-0.1 1-1.5 2.4 0.9 2.9-1.5 5.3-1.9 3.1-1.3 4.1-1.8 1.4-2.4-0.9-3.4 0.5-3.2-0.3-3 0.2-0.8 0.4 0.7 1.4-2.3-0.2-1.7 1.1-2 2.1 0 1.5 1.9 3-1.7 2-1.1 0-3.2-0.5-5.5 1.2-5.1-1 0.5-1 0.2-3.1-0.6-1.7-1.3-1.1-3.1-3.8-1.2-2.3 2.4 0.9 1-0.5-1.8-2.6 2.2-0.1 0.7-1-0.6-1.5-1.9-0.6-1.9-2.4-1.8-1.2-3.5-4.7-1.3-3.2-1.1 0.3-1.2-3.7-1.7-0.6-0.6-3.7-1.8-0.4-1.3-1.7-0.4-3.2-1.2-0.6-0.8-1.4-0.8-3-0.3-2.7-0.9-1.8 0.5-1.2 1.1-0.2 1.1 0.8 1 1.8 1.9-0.2 0.7-1.4 0.6-3.9-1.6-4.4 1.5-1.5 0.9-2.4 2.6-0.7 1.3-0.8 1.9-2 0.1-2.8 0.6-2.1-0.5-1.4-3.1-4.9-0.3-1.6 1.8-0.6 2.4 0 1.3-2.5 0.4-1.9-3.4-2.7-1.2-0.4-3.8-2.1 0.9-6.4 0-1.6-2.5-4.4 0.1-1.8-0.6-2.8 1.1-1.1-2.7-4.1 2-2.8-0.5-1.5 1.2-1 3.5-3.8 3.3-1.5 3.3-0.4 7.1 0.9 1.7-2.2-0.1-1.2-1.1-2.9-4.7-1.6 4.4-5 3.3-5 0.1-4.7 0.6-2.7-1.4-3.9 3.8-0.4 2.4-0.5 3.3-1.5-1-2.5 1.3-0.9 3-2.9 3.3-2.9 1.5-1 0.2-1.4-1-1.3-2.7-2.3 0.5-1.1 2.7-0.6 1.1-1.1 1.5-3.5 3.7-1.9 1.5-0.9 6.3 1.8 2.1-3.1-0.8-3.6 0-0.8 2.1-0.5 4.1 0.9 2.9 0.1 8.6 1.7 1 0.1 2.5-1.6-2.9-0.9 1.6-0.9 1.4-1.9-0.1-2-2.6-1.6 5.1-0.2 2.9 0.8 0.4 0.9 2.9 1.1 4.8 2.3 4.2 1.8 8.3 1.7 3.5 1.8 1.2 1.6 1.4 0.1 1.9 1.5 1.8 0.9-1.1 1 0.4 2.4 0.7 1.4-0.3 1.9 2.8 0.4 0.6 1.3-1 0.8-0.1 1.1 0.8 1.6 3.3 2.6 0.6 0.9-0.7 1.4 0.1 1.8-1.5 1.3 0 1.3 1.3 2.6 1.8 0.6 1.9 2.2 1.6 2.5z"
+                id="SE" name="Sweden">
+            </path>
+
+            <path d="M733.4 565.7l-2.9 0.5-0.1-1.5 2.2-0.6 0.8 1.6z m194.5-21.5l4.3 1.3 1.1-1 3.6 0.1 3.4 0.8 0.9-1.9 1.4-0.4 1.4 0.3 0.6 1.1 1.2 0.4 3.5 2.5 1.1 1.7 1.9 0.2 1.3 1.2 1.3 1.6 1.7 1.2 1.7 3.3-0.2 2.3-1 1.4 0.8 1.9 1.9 3-0.5 0.6 1.8 0.9 2.9 0.9 3.5-0.4 1.6 0.7 2.3 1.7 3.2 2.8 0.8 0.9-3.3-1.9-1.4 1.6-0.1 3.8-0.6 0.4-2.8-0.1-1.2 0.4 1.2 2.5 1.4 1.2 0 2.3 2.1 2.5 0.2 2.1 0.9 2.2 0.3 2.2 0.5 0.7 1.4 0.3-0.6 3.8-0.6 1.2-0.4 2 1.7 0.6 3.2 2 0.1 1.3 0.6 1.5 0.3 2 3.2 2.3 0 2.3-2.1-0.5-3.7 3.1-1.2-1.2-0.3-2.4-1.5-1.5-1.3-0.2-2.1 1.4-4.2-0.3-2.9-1.1-2.9-0.8-1.9 0.6-2.2-0.7-1.5 1.9-2 1.9-1.2 0.3-1.6-2.6-1.3-0.2-1.2 1.3-2.2 0.8-4.6 1-3.2 0.3-4.1-0.6-3.2 0.2-3.1 1.4-5.2 2.8-4.1 1.4-4.2 0.9-7.6-0.2-1.1-0.3-3.5-2.3-3.4-0.6-3.4 1.6-2.3 0.8-3.5 1.6-4.8-0.1-1.9-1.7-3.7-0.7-0.6 1.6-0.3 3.8 1.8 3.1-3.3 1.4 0 2.6-1.8 1.1-0.8 1.6-2.2-1.2-1-0.1 0.7-1.3-2.5-4.9 0.7-1.5 3.5-3.9-0.4-2.4-2-1.6-2.9 2.1-1.7 0.8-0.3 1.2-1.1 0.8-1.8 0.4-2.9-1-3.2-1.4-1.9-1.1-1.4-0.3-1.3 0.5-3.5 2.8-3.1 4.1-0.8 0.8-3.2 1.7-2.2 0.6-1-0.1-4.4 0.8-2.2 0.1-1.6 0.9-3.5-1-2.2-1.3-1.4-1.3-2.3-2.9-1.6-1.3-3.3-1.2-5.8-3-5.3-0.7-4-0.3-0.7 1.1 0.1 4.2-0.5 1.2-0.1 2.2-1.1 1.1-2.1-1-1.9 0.9-3.7 1.3-1.2 0.1-4.6-1.6-2.9-2.1-0.6-2-1.3-2.7-1.8 0.3-2.3-0.4-3.2-1.6-2.4-0.1-1.2 1.9-1 0.7-1.3-0.4 0.8-1.3-3.7 0.2-1.8 1-1.5-0.1-1-1 2-0.6 3.9-0.4 0.9-0.4 0.9-1.3 1.7-1.2-1.2-0.5-6 0.3-4.2-0.2-1 0.7-0.4-1.6 3.6-1.4-0.3-1.3-4.2-2.2-0.3-1.6-2-3.1 1.9-0.5 0.1-2.5-0.4-1.5-3.9-1.3-0.8 0.1-1.1-1.3-1.7-0.9-1.2 0.7-3.3-1.4-0.6-0.6 0.5-1.4 1-1.1-0.9-1.9-0.1-1 1.8-0.1 1 1.1 0.3 2.2 3.6 0.7 2.9-0.6-2.5-1.1-1.1-1.2-1.2-2.3 1.8-1.1 1.2-1.7-2.7-1.2 0.2-2.4-2.5-2.7 0.3-0.8 2.3-2.8-1.2-0.7-4.3 0.7-1.6 0.6-3.1 0.2-0.3-1.6 0.6-1.3-0.5-3.4 0.2-1.8 1.7-0.6 1.7-2.7 3-3.1 3.5 0 1.3-0.9 2.1 0 0.8 1.2 1.9 0.9 3.2-0.1 1.4-0.9-1.6-1.5 1.7-0.5 1.5 0.4-0.1 1.8 4.1-0.5 4.4 0.4 1.4-0.2 3.4 0 0.5-0.5-3.6-1.7 2-1.5 6.8-1.2 3.7-0.8-5.7-0.8-1.4-0.6-2-1.4-1-1.1 0.3-2.6 0.6-0.7 2.2-0.1 7.7 1.2 5.3-0.8 6.1 1.8 5.6-0.4 1.1-0.7 1.1-2.6 7.4-4.2 2.4-2.1 2.8-1.2 5-1.4 4-1.7 1.2-0.2 10.4 0.8 7 0.1 3-1.6 2 0.5-0.1 2.1 2.6 2.6 3.5 1.4 4.4-1.2 1.8 0.5 2.2 4 1.5 1.4 1.8 0.9 1.3 0.2 1.5-1.4 1.7-0.1 2.9 1.3 1.2 1.4 4.8 1.1 4.3 0.6 2 1.2 6.2 1.2 2.2-0.2 3.6-1.3 7-1.3 5.2 1.9 2.5 0.1 1.7 0.5 1.7-0.3 5-2.3 1.4-1.2 1.7-0.4 5.2-3.3 0.9-1.4z m-171.9-7.1l-0.1 1.8 1 2 2.3 2.7 2 1.4 8.1 3.3 1.4 0.3-0.9 3.5-2.1 0.5-6.4-1.5-1.6-0.1-2.9 1.4-2.3-0.4-3.1 0.6-0.6 2.1-2 2.4-3.3 1.9-2.5 1-3.5 3.7-1.5 2.2-1.5 0.7 0.5-2-0.3-1.7 2.3-2 3.3-1.6 0.7-1.3-5.4 0.3-3.3 0-1-1.8 0.8-0.4 2.5-3.2-0.5-4.4 3.4-2.5-0.9-3.1-2.3-1.5-1.2-0.3 0.3-1.4 1.9-0.3 1-2.2 3.4-0.4 1.5-1.1 3.5-0.3 3.4 2.5 1.5-0.6 1.8 0 0.9-0.4 1.7 0.2z"
+                id="TR" name="Turkey">
+            </path>
+
+            <path d="M796.7 471.6l0 0.6-3.9-0.8-0.9-0.9 4.8 1.1z m-26.4 14.6l-1.9-1.8-2.2-0.7-2.2 0.2-2.2 1.3-1.6 0.5-0.9-0.3-0.6 1.2-4-0.9-1.8-0.9-1.5-1.6 1.1-0.7 2-0.2 0.4-1-0.5-1.4 2.5-2.9 0-1.3 2.3-1.7 0.4-2-1-1.7-0.5-2.1 2.5-1.6 2.1 1 1.9 0.2 2.1 0.8 1.6-0.6 1.2 1.3 2.9-0.4 0.6-0.7-2.9-1.7-0.2-2.9-1-1.6-2.1-0.8-2-1.3-0.9-1.9 0-2.4-2.9-2-2.3-1.2-0.6-1.2 0.6-3.1-0.6-2.3-1.4-1.4-1.4 0.7-1.4-1-1.4-1.9-5.6-0.6-0.6-1.2-2.5-0.3-3.6-2.4-3.5-1-1.6 1-0.8-0.3-1.2 1-4-0.1-0.7-0.2-2.3 1.4-0.2 0.6-3.7 0.8-0.2 1.4-1.1 1.9-8.4 1.2-3.4 1.3-1 1.2-2.1 0.5-4.3-3.3-1.2-0.3-3.6 0.7-6.9-1.4-2.7 0.1-2.7-1.5-0.7 0-0.9 1.2-2.1 0.9-0.7-1.7-0.9-0.7-2.3-0.4-0.9-1.1-2.2-0.8-1.4-2.3-1.5 0-0.1-2.5 1.6-1.7 0.8-2.9 1.5-3 3.3 0.7 0.4-0.9-1.9-1.4 0.1-1.8-1.5-3.8 0.6-1 3.4-4.4 2.4-2.7 4.8-4.6 3-0.5 1.2-1.8-0.1-1.3-1.8-2.6 1.2-1.3-1.4-1-1.9-2.8-2.7-2.8 0-1.3-1.2-1.8-0.1-1.4 1.1-0.5 1.9 0.3 1.1 0.5 1.9-1.2 1.6-1.6 0.8-1.4 5.9-0.5 2.3-0.5 2.4-0.1 7.8 0.4 6.5 1.5 6.1 0.9 2.3 0 1.8 2.1 1 0.3 3-0.1 1.1 0.4 0.8 1.4 1.2-1.7 1.9 0.5 1.9-0.7 5.1 1 0.8 1.3 1 0.3 0.6-1.1 3.8-1.6 2.5 3 0.6 0.5 2.3-0.8 7.1-0.6 2 1.1 0.5 1.2 2.7 0.9 0.8-1.4-1.8-3.6 1-3.3 0.8-1.2 2.2-2.3 1.1-0.5 1.7 0.4 1.4-0.8 5 0.1 2.6 0.8 1.7-0.1 1.8-0.9 0.5-2.4 1.6-0.6 3.7 0.9 1-0.1 2.7-1.2 1.6-0.2 2 0.3 3.4-0.2 2.5 1.4 1.4 1.4 1.7 2.7 4 3-0.1 1-2.9 0.5 1.3 1.9 0.5 2 1.2 1.7-0.4 1 3.1 0.1 2.9 1 4.1-0.5 1.5 2.2 1.9 0.2 0.2 0.9 1.4 2.8 0.9 1.1-0.2 1.8 0.5 1 1.9 1.6 0.7 1 1.1 0.4 2.3-1.4 2.8 0.4 2.7 2 2.3 0 1.7 1.1 1.4-1.3 4.3-0.9 2.3-1 1 0.1 2.3 1.8 0.5 1.2 1.5 1.6 4.7 2.9 1.3-0.6 0.9-1.5 2.6 1.4 2.4 0.2 3.6 1.9 1.4 0.1 1.6-0.6 2 1.8 1.9 0.2 4.4 2.3 2.9-0.3 1 0.8-0.1 1.7 1 0.9-0.1 2.6-1.7 2.1-2.3 0.8 1.1 1.5 3.1 0.9-0.9 0.9-1.4-0.2-0.8 1-0.3 2.3 2.5 0.7 1.2 2.8 0 1.1 1.2 0.3-1.3 4.7-0.1 1.8-4.2 0.2-6-0.3-0.9 0.2-0.9 1.9-2.3 1.4-1.7 0.2-0.8 0.8 0 2.3-0.4 1.3 1.1 0.7-0.7 1.6 0.5 1.3-0.5 0.1-4.3-0.2-3.4 0.3-2.1 2.4-1.5 0-3.2 1.5-1.4 1.7-1.4-0.8-3.1 0.5-1.7 1.1-3.1-0.1-2.3 0.6-4.6 3.8-1.3 0.1-1-1-1.6 2.6-2.4 1.2 0.6 3.1 2.8 4.5 3.6 3.9 1.7 1.4 2.4 0.7 3.1-1.4 2.4 0.4 1.7-1.2 1.5-0.1 3.7 1-1.1 2.5-0.3 2.6-1.9 0.6-2.2-0.1-2.2 0.4-2.7-1.5-2.5 0-1.1 1.9-2.3 1.3-0.6 1.4-2.6-0.3-2.1 0.2-2.8 1.4-1.9 2.8-2.3 1.8-1.9 0.5-2-0.1-1.3-0.6-2.9-1.8 0.7-2 0.5-3.5-1.2-2.9-2.3-1.4-1.6 0.2-0.9-0.3-3.7-2.4-1.9-0.2-1.9 0.5-1.5-1.2 3.5-2.9 3.5-2.4 1.7-0.2 2.1-1.2 2.2-1.6-1.3-2.3-1.9 0.6-2.2-1.1-0.9-0.8-3.1 0.8-1.8-0.1-3.8 0.8-1.9-0.8-4-2-2.6-0.3 0-1 1.8-0.3 0.1-1-3.8-0.6-0.1-1.3 2 0.5 3.2 0.2 2.9 0.5 2.1-1.8-2.4 0.5-2.9-0.5-2.1-1.8-2.2 1.2-3.2-0.5-2.4 0.8-2.3-0.2-4.1 1.1-0.4 2.9-2 4-3.1 3.7-4.6 2.1-1.7-0.4-0.9 1.1 0.2 1.3 1 0.9 1.1 3.1-0.1 1.3z"
+                id="UA" name="Ukraine">
+            </path>
+
+            <path d="M674 532.8l-2.1 0.4-1.5 1.9-1.6-1.2-1 0.1-3.4 1.5-0.2 2.6-0.5 0.8-1.7-0.3-0.1-2.2-1.4-3.2-1.9-1.3-1.4-0.5-2-2.9-0.5-0.4 0.2-1.3-0.9-1.5 2-0.4 0.2-0.7 1.6-0.4 1.5-0.5 0-1.4 1.7-1.3 0.4-1-0.6-1.6 1.7-0.9 1.1 1.3 2.8 1.3 2.3 2.1 0.2 0.7 2.2 1.2 0 1.3 4.7 1.2-0.2 1.2-1.2 3.2-1 0.9 0.6 1.3z"
+                id="XK" name="Kosovo">
+            </path>
+
+            <path d="M452.2 392.5l-0.6 1.2-3.2 1.5-3.9-1.1-1.2 0.6-1.2-0.7-0.3-1.3 2.8-0.4 3.5 0.7 1.5-0.6 1.2 0.5 1.4-0.4z m-3.4-5.3l1.5 1.4-1.5 0.3-1.6-1 1.6-0.7z m18.1-8.5l-3-0.4-0.9-0.7 2.4-1.7 3.2-1.5 1.7 0.4 0.5 0.9-1.1 1.1-2.8 1.9z m-7.6-11.5l-1.1 1.1-0.6-1.4 1.7-1.4 0 1.7z m-7.1 25.3l-2.6-0.8-2.2 0.5-2.8-0.7-1.7-1.3 1.2-0.7 2.3-0.1 4 0.7 1.2-0.9-0.6-1-2.3-2.1 1.6-2.7 1.4-1 3.2-3.8 0.9-2 1.2-5.5 1-2 1.4 0.5 2-0.8 3.4-2 1-1.7 1-0.9 4-1.6 2.1-0.4 8.9-0.5 1.8 1.7 2.7 0.6 0.1 4.3-0.7 1.7-1.1 3.8-3.1-0.1-0.6 1.6 0.7 1 2.6 0.3 0.8 1-0.6 2.6-2.6 1.6-0.3 0.9 1 0.9-0.7 1-4.5 1.3-2.3-0.8-1.8 0.7-0.7 1.1 3.1 4.1 0 1.2-1.3 2.8 0.7 1.2-2.2 1.6-0.5 0.8 1.9 1.2-0.6 2.4-1.2-0.1-2.4-0.3-0.7-1.1 1.1-1.3 1-2.9-2.6-1.1-1.7-1.3-1.9 0.4-2.6-1.4-0.5-1.6-2.5 0.8-1.1-1.3-1.5 1-1.5-0.8-1.4 0.7 0.2 0.9-2.1-0.3z m18.1-15.6l0.9-0.9-0.5-1.9-2.1-0.3-0.5-0.8 0.4-2.5-2.8-0.4-0.4-2.5-1.5 0.3-2.3 1.5 0.6 1.3 1.6 1.6 0 0.9-2.1 0.7 0.1 2.6-0.2 1.6 5.4 1.4 3.4-2.6z"
+                id="NL" name="Netherlands">
+            </path>
+
+            <path d="M420.4 584.7l-0.4 1-2-0.4 0-1.5 1.2-1.6 2.7-0.7 0.8 1.3-2.3 1.9z m21.7-13.6l1.2 0.5 1.3-0.5 1.4 0.4 0.2 1-2.2 3.5-0.6 1.4-2.1 1.3-4-1.7-0.3-1.6-1.4-0.7-1.7 1.3-1.4-1-0.3-1.2 5.3-3.7 1.5-0.9 3.2-0.9 0.5 1.2-0.6 1.6z m14.7-0.8l-0.3 0.2-3.9-1.9-1.6-0.4 0.1-1.4 2.6-0.2 2.1 0.7 1.2 1.8-0.2 1.2z m-77.4-55.5l0.5 1.3 4.3 1.3-0.3 3.2 1.8-0.4 4.4 1.7 2.2 0.2 2.2 2.1 3.6-0.4 3.2 2.1 3.1-0.4 0.7 0.4 4.7 0 0.8-2.4 4 1.1 2.5 1 1.7 0.3 1.1 1.5 0 2.1 0.7 0.4 2.8-1.1 2.8 1.2 1.3 1.2 2.1-1.1 4.7 1.2 1.2-0.8 3.8-1.1 3.1 0.6 1.2 2.2-1.8 0.5-0.2 1.4 1.2 1.3 0.1 2.1-3 2.8-8.7 4.7-2.9 2.7-6.6 1.5-4.5 1.5-2.2 0.5-4.1 3.8 1.1 0.3 1.2 1.2-0.4 0.6-2.5 1.1-0.9 0-2.9 4.6-2.6 3.3-1.5 1.5-1.5 2.1-3.2 5.6-0.1 1.6 1.6 5.5 2.2 2.7 2.4 1.1 0.6 1-0.8 1-2.4 1.7-4.3 2.3-1.8 1.9-0.3 1.8-1.3 0.8-0.5 2.5-1.8 4.4 1.3 1.2-1.3 0.8-6.5 0.3-4.1 2.8-2 2.4-1.9 4.5-2.2 2.6-1 0.5-1.5-1.2-1.9-0.1-1.9 0.3-1 1-1.5 0.5-1.5-0.5-4.6-0.2-2.3 0.8-1.9-0.5-3.2-0.3-7 0.6-4 3.4-3.4 0-3.1 1.2-2.1 3-0.5 1.6-1-0.3-0.3 1.2-2.1 0.8-2.4-1-2-1.5-1-0.1-1.6-2.2-1.2-3.1 0-1.1-1.5-0.6-0.3-1.4 0.2-1.6-1.2-2-5-3.7-3.7-0.3-3 0.5-1.1-6.4 0.8-2.2 3.5-4.3 2.1-0.6 0.7-2.4-1.7 0.2-3-4.3 0.9-4 3.2-3 0.6-2.4-0.5-0.8-1.7-0.4-1.6-3.2-0.3-1.9-1.4-1.2-1.1-1.9 1.1-0.3 4.3 0 1-0.5 1.7-3.5 0-1.9-1.4-1.3 0.2-1 2.7-2.1-0.4-1.1 0.4-3.8-0.4-3.5-0.8-2 2-1.6 1.1-1.6 1.6-1.4 2.1-1.1 2.5-2.5-0.3-1.1-1.9-0.9-1.9-0.2-0.1-3.1-0.7-1.1-5.2-0.1-1.5-0.5-0.8 1.4-2.5 0.9-1.4 0-2.2-0.8-2.9 0.1-3 1.1-0.9-1.3 1.3-1.9-1-1.8-4.1 1-1.9 0.9-1.2 1.1-1.3-0.1 0-2.5 2.6-2.6-1.6-0.2 1.2-1.9-1-1 0.1-2.6-2.3 0.9-0.5-1.2 1.4-2.2-1.4-0.2-2.3-2.5 0.8-3.1 2-1 1.9-1.5 2.6 0.3 1.7-0.3 3.6-1.6-0.4-1.5 0.4-0.7 3.2-1.7 2-0.2 1.9-0.9 1.3 0.5 1.1-0.1 1.3 0.6 1.7 1.6 2.5 0.6 2-0.5 3.6-0.1 1.7 0.2 3.2-0.3 1.8 0.1 2.9-0.8 2.3 1 4.4 0.4 2.6 0.8 7.2 1.4 2.7 0 3.7-0.8 1.6-0.5 1.5 0.3 2.1-0.6 2.3 1 4.7 1.3 2.1-1.3 3.4 0.6 3.3 1.3 1.8 0.1 2.6-0.3 2.5-1z"
+                id="ES" name="Spain">
+            </path>
+
+            <path d="M521.2 524.1l-0.3 2.3 1 1.7 0.6 6.5-1.8 3.2 0.1 3-1.6 4.3-0.8 1.1-3.8-2-1.2-1.1 1-1.8-2.2-0.9 0.4-1.7-0.2-0.9-1.6-0.5 1-1.3 0-0.8-1.5-1-0.3-0.9 1.3-1.1-0.6-0.9 0.1-1.3 0.9-1.9 1.2-0.9 2.8-0.8 1.2-1.1 1.9 0.6 0.6-1.2-0.2-2.7 0.3-1.1 1.3 0.5 0.4 2.7z m-133.9-47.9l-0.4 1.4-1.9-2.4 1-0.5 1.3 1.5z m84.4-55.6l2.1 1.2 1.4 0 2-0.6 1.3 0.7 1.4 0.1 3.5 4.3 1.8-0.7 1.4 0.4 0.3 0.9 2.1 0.1 3-0.6 2 1.4 4.8 0.8 1.6 0.6 0.1 1.3-3.5 3.8-1.4 5.5-1 1.8-0.3 1.5 0.2 2.4-0.6 2 0 1.5 1.1 1.3-1.5 0.7-0.8 1.4-2.7 0.3-0.8-0.8-1.1 0-1.8 1.4 1.3 1.1-0.6 0.9-3.8 4-2.1 0.9-0.5 2.9-3 2.2-1.1 2.9 0.8 0.7-0.3 1.5-1.5 1 0 1 1.4 0.1 2.3-1.6-0.5-1.3 2.3-1.5 1.9-0.1 2.2 0.3 0.8 2.2-0.5 1.7 2.3 2.3 0.9 1.4-2.6 1.7 0.1 1.6 0.9 0.6 1.7 2.6 1.8 1.6-0.8 2.5-1.2 0.3-1.7 1.3-1.9-0.2-0.8 0.4 1.5 3.1 3.2 1.4 0.5 1.7-0.8 0.7-1.4 2.5 0.7 1.3 0.1 1.4 0.9 0.9 5.1 2.4 3.3-0.6 0.3 2-2.1 2.6 0.2 1.5-0.6 0.3-0.8 0.3-2.4 1.1-3.9 3.4-1.8 1-0.7 1.8-2 1.4-2.4 0.5-2.3 1-1.1-0.4-2.8 0-1.7-1.2-3.4-0.8-1.1-1.8-2.5-0.1-0.8-1.5-1.9 0.3-1.5 0.7-2-0.1-5.9-1.8-1.4-1.4-1.7 0.5-1.6 1.6-6.5 4.1-2.6 4.3-0.1 1.2 0.7 3.9 1.6 2.5-3.1-0.6-3.8 1.1-1.2 0.8-4.7-1.2-2.1 1.1-1.3-1.2-2.8-1.2 0-1.5-2.6-0.6-0.9 0.7-1.1-1.5-1.7-0.3-2.5-1-4-1.1-0.8 2.4-4.7 0-0.7-0.4-3.1 0.4-3.2-2.1-3.6 0.4-2.2-2.1-2.2-0.2-4.4-1.7-1.8 0.4 0.3-3.2-4.3-1.3-0.5-1.3 2-0.5 1.9-1.9 1.7-7.1 1.3-8.3 1-1.6 1.2-0.4-1-1.2-1.1 1.5 0.7-7.6 0.5-2.8 0.9-3 3.1 2.4 0.8 1 0.9 3.4 0.6 0.1-1.1-4.6-1.8-2.3-3.6-2.3-0.5-1.3 1.8-0.6-0.5-1.8-0.5-5.9-3-0.6-4.8-2.6-1.7-2.6-1.6-1.9-0.4-1.7 0.9-1.8-0.8-1.1-1.4-0.8 1.1-1.6-3-0.2-1.8-0.5-0.3-1.1 1.3-1.4-1.6-0.8-2.6 0.2-0.2-1.7-2.2 0.3-1.2-0.3-1.1-1.1-1.3 0.2-2.2-0.4-0.8-0.7-4.8-1.3-2-0.1-1.9 0.6-1.1-0.3-1.3-2.2-3.1-1.1 0.6-0.7 3.1-0.6 0.6-0.7-2.2-0.9-0.2-1.2 1.6 0.3 0.1-1.2-4 0.1-0.4-1.3 0.5-1.4 2.3-1.2 5.8-1.4 2.5 0.2 1.7-0.2 3-1.6 2.9-0.4 2.7 0.8 2.5 2.8 1.2 1 3.1-1.7 4.4 0.1 0.9 0.9 1.2-1.7 1 1 4.7-0.2-1.5-2.5-0.2-6-1.3-1.7-2.1-4.3 0.2-1.4 3.2 0.3 2.7-0.6 1.3 0.4 0.3 2.8 1.1 1.6 2.2 0 2.3 0.5 3 0.1 4.3 0.8 1.8-0.5 3.5-2-1.8-0.7-0.2-0.7 0.9-2.2 5.2-2.5 3.7-0.7 3.8-1.3 2-1.4 1.2-1.8 0.5-1.3 0.3-6.8 1-2.2 2.9-1.6 6.4-1.1 0.9-0.5 0.9 2.2 0 1.2 2.9 2.4 3.2-1 1.6 1.8 0.5 2 4.2 1.1 0.6 2 0.9-0.3 3.6 0.4 1.6 1.1-0.4 1.6 0.7 0.7-0.7 1.4 0.5 0.6 2.3 0.3 2.2-0.3 1.3-0.6 0.5-1.4 1.9-0.6-0.8 2.7 0.6 0.6 0.4 2 1.7 0.1 3.3 1.5 2.8 2.6 3.4-0.4z"
+                id="FR" name="France">
+            </path>
+
+            <path d="M306 611.9l-1.2 0.2-4.5 2.5-1.3 0-2.5-1.1-4.5-0.4-1.5-0.3-1.8 0.7-1.5 0-1.1 1-0.8-0.3 1-2.1 1.6-4.1 0-2.6 0.4-2.2-0.3-2.1-0.7-1.4 1.1-3.5 0-1.8-0.9-2.3 2.8 0.3-1.7-1.4-1.5 0-2.3 0.9-1.5 0.1 0.1-1.4-0.5-1.8 2-0.6 1.6-1.7-0.3-1.6-1.2 1.1-0.6 1.4-2.8 0.7-1.5-0.5 0-1.9 0.6-1.7 0.4-2.3 0.8-2.2-0.2-1.4 1.6-1.3 1.4-1.9 4.3-8.9-0.6-1 1.7-6.6 1.2-2.3 0.6-4.3-0.2-2-0.7-2-0.7-4.3 0-1.4-0.9-1.8 0.1-1 1.4-1.7 1.2-1.1 1.9-0.9 4.1-1 1 1.8-1.3 1.9 0.9 1.3 3-1.1 2.9-0.1 2.2 0.8 1.4 0 2.5-0.9 0.8-1.4 1.5 0.5 5.2 0.1 0.7 1.1 0.1 3.1 1.9 0.2 1.9 0.9 0.3 1.1-2.5 2.5-2.1 1.1-1.6 1.4-1.1 1.6-2 1.6 0.8 2 0.4 3.5-0.4 3.8 0.4 1.1-2.7 2.1-0.2 1 1.4 1.3 0 1.9-1.7 3.5-1 0.5-4.3 0-1.1 0.3 1.1 1.9 1.4 1.2 0.3 1.9 1.6 3.2 1.7 0.4 0.5 0.8-0.6 2.4-3.2 3-0.9 4 3 4.3 1.7-0.2-0.7 2.4-2.1 0.6-3.5 4.3-0.8 2.2 1.1 6.4z"
+                id="PT" name="Portugal">
+            </path>
+
+            <path d="M843.2 646.3l-0.1 0-0.1 0 0.1-0.1 0.1 0.1z m0 0l0.1 0 0.1 0 0.1 0.1-0.1 0-0.1 0-0.1 0 0-0.1z m-0.4-0.3l-0.1 0 0-0.1 0.1 0 0-0.1 0.1 0.1 0 0.1-0.1 0z m3.5-0.8l0.2 0.1 0.3 0.3 0.2 0.4 0.1 0.3 0.2 0.2 0.2 0.1-0.2 0-0.3 0-0.3-0.1-0.2-0.2-0.2 0-0.7 0.1-0.6 0.1-0.1 0.1-0.1 0.1-0.3 0-0.3 0-0.1-0.1 0.2-0.1 0.1 0-0.1-0.3-0.2 0-0.2-0.1 0-0.4-0.1-0.1-0.3-0.1 0-0.2 0.1-0.1 0.2-0.1 0.1 0.1 0.1 0.1 0.2-0.1 0.1-0.1 0.1-0.1 0.1-0.1 0.2 0.2 0.2 0.1 0.1 0 0.2 0.1 0.1 0 0.3-0.1 0.2 0 0.1 0.1 0.2 0 0.1 0 0.1 0 0-0.1z m-4 1.1l0 0.2-0.3 0.1-0.3 0.3 0 0.2-0.1 0.3 0 0.3 0 0.3 0 0.3 0 0.1-0.1 0-0.1 0.1 0 0.3-0.1 0.1-0.2 0-0.4 0-0.2 0.1-0.1 0.1-0.1 0.2-0.1 0.1-0.2 0.1-0.4 0.1-0.4 0.2-0.3 0.1-0.2-0.3 0.1 0.5-0.4 0.2-0.9 0.2-0.2 0.1-0.1 0.1-0.2 0.1-0.2 0-0.9-0.1-1 0.1-0.1 0.1-0.2 0.1-0.6 0.5-0.1 0.3-0.3 0-0.3 0.1 0 0.1-0.1-0.1-0.1-0.2 0.1-0.2 0.1 0.1 0.3 0-0.1-0.4-0.4-0.1-0.2 0.2-0.1 0.1-0.2-0.1 0 0.1-0.4-0.1-0.3-0.2-0.1-0.1 0-0.2-0.1 0-0.1 0-0.1 0-0.1 0-0.1 0.1 0 0.1-0.2 0-0.1 0.2-0.2 0-0.2 0-0.1 0.1 0.1 0.2-0.2 0.1-0.5 0-0.6-0.1-0.5-0.2-0.7-0.4-0.3-0.1-0.3-0.1-0.5 0-0.3-0.1-0.2-0.2-0.6-0.4-0.1-0.1-0.1-0.4 0-0.1-0.1-0.1-0.4-0.9-0.1-0.1-0.3-0.2-0.2-0.1-0.1-0.1-0.1-0.2 0-0.2 0.1-0.2-0.3-0.4-0.1-0.2 0-0.4-0.1-0.2-0.3-0.5-0.2-0.3-0.1-0.4 0.1-0.3 0.1 0 0.3 0.1 0.4 0.5 0.3 0.1 0.2 0.1 0.6 0 0.4-0.2 0.3-0.4 0.3-0.4 0.1-0.5 0.1-0.3 0.2-0.1 0.1-0.1 0-0.1 0.1 0 0.1 0.1 0.1 0 0.2 0 0.1 0.1 0.2 0.2 0.2 0 0.1 0 0.2-0.2 0.1-0.2 0-0.2 0.4 0.1 0.2 0.4 0.1 0.2 0.1 0.2 0.2 0.1 0.4 0.1 0.3 0.1 0.2 0.2 0.2 0.1 0.1 0.2 0 0.2 0.1 0.1 0.1 0.1 0.1 0 0.4-0.1 0.1-0.2 0.1-0.2 0.1-0.1 0.1 0.1 0.1 0.1 0.2 0.1 0.2 0 0.2-0.1 0.2-0.2 0.2-0.2 0.2-0.1 0.3-0.2 0.1-0.2 0.2 0 0.3 0.1 0.3 0 0.4 0 0.1-0.1 0.1 0 0.1-0.2 0.3-0.2 0.2 0 0.2 0.1 0.1 0.1 0.2 0 0.1 0 0.2 0.2 0.2 0.2 0.2 0.1 0.3 0.1 0.1 0 0.1-0.1 0-0.1 0.1-0.2 0.3-0.3 0.1 0 0.2 0 0.3 0 0.2 0 0.1 0 0 0.2 0.1 0.2 0.1 0.1 0.2 0.2 0.2 0.2 0 0.1 0.1 0.3 0 0.2 0 0.3 0 0.4-0.1 0.3 0.1 0.2 0.1 0.1 0.2 0 0.2 0 0-0.3 0-0.4 0.1-0.1 0.1-0.1 0.1-0.1 0.1 0.1 0.1 0.4 0.1 0.2 0.1 0.1 0.3-0.1 0-0.2 0.1-0.1 0.2 0 0.1 0.1 0.2 0 0.1-0.1 0.1 0.1 0.1 0 0.3 0.2 0.2 0 0.1 0 0.3 0.2z"
+                id="CY" name="Cyprus">
+            </path>
+
+            <circle cx="399.9" cy="390.8" id="0">
+            </circle>
+            <circle cx="575.4" cy="412" id="1">
+            </circle>
+            <circle cx="521" cy="266.6" id="2">
+            </circle>
+        </svg>
+    );
+}
+export default MapViewer;
